% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrl.R
\name{contrl}
\alias{contrl}
\title{Conspecific trivariate random labeling, tidy data}
\usage{
contrl(ppp.marks,yes="Yes",no="No",...,
    r = NULL,
    kernel = "epanechnikov", bw = NULL, stoyan = 0.15,
    correction = c("isotropic", "Ripley", "translate"),
    divisor = c("r", "d"),ratio = FALSE)
}
\arguments{
\item{ppp.marks}{A multitype "\verb{ppp}" object, of which the binary levels are
yes and no.}

\item{yes}{The value of successful category in the dichotomous mark,
for example, "Yes". If \verb{missing}, the second level of the mark will be assigned.}

\item{no}{The value of failure category in the dichotomous mark,
for example, "No". If \verb{missing}, the first level of the mark will be assigned.}

\item{...}{Same as in \verb{pcfdot()} from the package \verb{spatstat}}

\item{r}{Same as in \verb{pcfdot()}from the package \verb{spatstat}}

\item{kernel}{Same as in \verb{pcfdot()} from the package \verb{spatstat}}

\item{bw}{Same as in \verb{pcfdot()} from the package \verb{spatstat}}

\item{stoyan}{Same as in \verb{pcfdot()} from the package \verb{spatstat}}

\item{correction}{Same as in \verb{pcfdot()} from the package \verb{spatstat}}

\item{divisor}{Same as in \verb{pcfdot()} from the package \verb{spatstat}}

\item{ratio}{Same as in \verb{pcfdot()} from the package \verb{spatstat}}
}
\value{
An object of class "\verb{fv}".
}
\description{
Trivariate random labeling function for one variate with a binary qualitative mark.
}
\details{
For example, this function tests the influence of the spatial pattern of
conspecific host tree on infection (Raventos et al. 2010).
\deqn{p_{l,l+m}(r)=p_l\frac{g_{l,l+m}(r)}{g_{l+m,l+m}(r)}}
where \eqn{p_l} is the proportions of type \emph{l} points in the pattern,
\eqn{g()} is the  pair-correlation function, \emph{m} is another type points.

Performs the same job as \code{\link{contrl2}}, but for tidy data, i.e, with only one mark.

For envelope simulation, let \verb{simulate=NULL} in the
function \verb{envelope} from the package \verb{spatstat} to generate simulations of
Complete Spatial Randomness (CSR). Alternatively,
use \verb{rlabel} from the package \verb{spatstat} to permutate labels.
}
\examples{
\dontrun{
nsim<-2499
plot(contrl(ppp.ash.c3,yes="Yes",no="No"))
ev.contrl.ash.c3<-envelope(ppp.ash.c3,contrl,nsim=nsim,
    funargs=list(yes="Yes",no="No"),
    simulate=rlabel)
plot(ev.contrl.ash.c3)
}

}
\references{
Wiegand, T. and Moloney, K. A. 2013. Handbook of spatial point-pattern
analysis in ecology. Chapman and Hall CRC, Boca Raton.

Raventos, J., Wiegand, T., and de Luis, M. 2010. Evidence for the spatial
segregation hypothesis: a test with nine-year survivorship data in a
Mediterranean shrubland. Ecology. 91 7:2110-2120.
}
\seealso{
\code{\link{contrl2}}, \code{\link{heterotrl}},\code{\link{trig_g}}.
}
\author{
Tianzhong Jing \email{Jingtianzhong@163.com}
}
