% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_gcross.R
\name{g_gcross}
\alias{g_gcross}
\title{\verb{pcfcross} difference estimation, tidy data}
\usage{
g_gcross(ppp.marks,yes="Yes",no="No",t.type="yn-nn",...,
    r = NULL, kernel = "epanechnikov", bw = NULL, stoyan = 0.15,
    correction = c("isotropic", "Ripley", "translate"),
    divisor = c("r", "d"),ratio = FALSE)
}
\arguments{
\item{ppp.marks}{A "\verb{ppp}" object with a dichotomous mark(for example, "Yes" and "No").}

\item{yes}{The value of successful category in the dichotomous mark, for example, "Yes".
If \verb{missing}, the second level of the mark will be assigned.}

\item{no}{The value of failure category in the dichotomous mark, for example, "No".
If \verb{missing}, the first level of the mark will be assigned}

\item{t.type}{Character, indicating what type of test to be estimated. The value can be "yn-nn", "ny-nn", "yy-nn", and "ny-yn",see \verb{Details}}

\item{...}{Same as in \verb{pcfcross()} from the package \verb{spatstat}}

\item{r}{Same as in \verb{pcfcross()}from the package \verb{spatstat}}

\item{kernel}{Same as in \verb{pcfcross()} from the package \verb{spatstat}}

\item{bw}{Same as in \verb{pcfcross()} from the package \verb{spatstat}}

\item{stoyan}{Same as in \verb{pcfcross()} from the package \verb{spatstat}}

\item{correction}{Same as in \verb{pcfcross()} from the package \verb{spatstat}}

\item{divisor}{Same as in \verb{pcfcross()} from the package \verb{spatstat}}

\item{ratio}{Same as in \verb{pcfcross()} from the package \verb{spatstat}}
}
\value{
A object of "\verb{fv}".
}
\description{
Estimates difference between two cross-type pair correlation function (\verb{pcfcross}).
}
\details{
Bivariate difference pcf (or \verb{g} function) is estimated as (Raventos, et al., 2010).

if t.type=="yn-nn", \deqn{g_{m,l}(r)-g_{l,l}(r)}
if t.type=="ny-nn", \deqn{g_{l,m}(r)-g_{l,l}(r)}
if t.type=="yy-nn", \deqn{g_{m,m}(r)-g_{l,l}(r)}
if t.type=="ny-yn", \deqn{g_{l,l}(r)-g_{m,l}(r)}
where \eqn{g()} is the pair-correlation function (pcf), \emph{l} and \emph{m} are
two levels of a binary mark.

Note that there is a symmetry in \emph{i} and \emph{j}, so that \eqn{g_{ij}(r) = g_{ji}(r)},
However, estimates of these functions may not be exactly symmetric in i and j,
because of the effect of edge corrections.

This function tests, for example, the host density-dependent infection of a
forest disease.

Performs the same job as \code{\link{g_gcross2}}, but used for tidy data,
i.e, with only one mark.

Use the function \verb{rlabel()} from the package \verb{spatstat} to shuffle
the binary mark for envelope simulation.
}
\examples{
\dontrun{
plot(g_gcross(amacrine,yes="off",no="on",t.type="ny-nn"))
plot(g_gcross(ppp.ash.c3,t.type="ny-nn"))
nsim=2499
ev.g_gcross.ash.c3<-envelope(ppp.ash.c3,g_gcross,nsim=nsim,
      funargs=list(yes="Yes",no="No"),simulate=rlabel)
plot(ev.g_gcross.ash.c3,ylim=c(-2,2))
}

}
\references{
Wiegand, T. and Moloney, K. A. 2013. Handbook of spatial point-pattern
analysis in ecology. Chapman and Hall/CRC, Boca Raton.

Raventos, J., Wiegand, T., and de Luis, M. 2010. Evidence for the spatial
segregation hypothesis: a test with nine-year survivorship data in
a Mediterranean shrubland. Ecology. 91 7:2110-2120.

Baddeley A.,Rubak E.,Turner R. (2016). Spatial Point Patterns: Methodology and Applications with R.
Boca Raton, FL: CRC Press.
}
\seealso{
\code{\link{g_gcross2}} which does the same job,
\code{\link{K_Kcross}}, \code{\link{L_Lcross}}, \code{\link{J_Jcross}},\code{\link{ND_NDcross}},
\code{\link{trigcross}}, \code{\link{trig_g}}.
}
\author{
Tianzhong Jing \email{Jingtianzhong@163.com}
}
