% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_gdot.R
\name{g_gdot}
\alias{g_gdot}
\title{\verb{pcfdot} difference estimation, tidy data}
\usage{
g_gdot(ppp.marks,yes="Yes",no="No",...,
    r = NULL,
    kernel = "epanechnikov", bw = NULL, stoyan = 0.15,
    correction = c("isotropic", "Ripley", "translate"),
    divisor = c("r", "d"),ratio = FALSE)
}
\arguments{
\item{ppp.marks}{A "\verb{ppp}" object with a dichotomous mark(for example, "Yes" and "No").}

\item{yes}{The value of successful category in the dichotomous mark, for example, "Yes".
If \verb{missing}, the second level of the mark will be assigned.}

\item{no}{The value of failure category in the dichotomous mark, for example, "No".
If \verb{missing}, the first level of the mark will be assigned.}

\item{...}{Same as in \verb{pcfdot()} from the package \verb{spatstat}}

\item{r}{Same as in \verb{pcfdot()}from the package \verb{spatstat}}

\item{kernel}{Same as in \verb{pcfdot()} from the package \verb{spatstat}}

\item{bw}{Same as in \verb{pcfdot()} from the package \verb{spatstat}}

\item{stoyan}{Same as in \verb{pcfdot()} from the package \verb{spatstat}}

\item{correction}{Same as in \verb{pcfdot()} from the package \verb{spatstat}}

\item{divisor}{Same as in \verb{pcfdot()} from the package \verb{spatstat}}

\item{ratio}{Same as in \verb{pcfdot()} from the package \verb{spatstat}}
}
\value{
A object of "\verb{fv}".
}
\description{
Estimates difference between two dot-type pair correlation functions (\verb{pcfdot}).
Same as g_gdot2.
}
\details{
Bivariate difference pcfdot (or \emph{g} function) is estimated as (Raventós, et al., 2010).
\deqn{g_{l,l+m}(r)-g_{m,l+m}(r)}
where \eqn{g()} is the  pair-correlation function, \emph{l} and \emph{m} are
two levels of a binary mark.

Use the function \verb{rlabel()} from the package \verb{spatstat} to shuffle
the binary mark for envelope simulation.
}
\examples{
\dontrun{
plot(g_gdot(ppp.ash.c3),ylim=c(-3,1))
nsim=2499
ev.g_gdot.ash.c3<-envelope(ppp.ash.c3,g_gdot,nsim=nsim,
    funargs=list(yes="Yes",no="No"),simulate=rlabel)
plot(ev.g_gdot.ash.c3,ylim=c(-4,4))
}

}
\references{
Wiegand, T. and Moloney, K. A. 2013. Handbook of spatial point-pattern
analysis in ecology. Chapman and Hall/CRC, Boca Raton.

Raventós, J., Wiegand, T., and de Luis, M. 2010. Evidence for the spatial
segregation hypothesis: a test with nine-year survivorship data in a
Mediterranean shrubland. Ecology. 91 7:2110-2120.
}
\seealso{
\code{\link{g_gdot2}} which does the same job,
\code{\link{K_Kdot}}, \code{\link{L_Ldot}}, \code{\link{J_Jdot}}, \code{\link{ND_NDdot}},
\code{\link{trig_g}}, \code{\link{trigcross}}.
}
\author{
Tianzhong Jing \email{Jingtianzhong@163.com}
}
