% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_gdot2.R
\name{g_gdot2}
\alias{g_gdot2}
\title{\verb{pcfdot} difference estimation, untidy data}
\usage{
g_gdot2(ppp.marks,V.species="Species", host="Ash",
    V.bistate="Infected",yes="Yes",no="No",...,
    r = NULL,
    kernel = "epanechnikov", bw = NULL, stoyan = 0.15,
    correction = c("isotropic", "Ripley", "translate"),
    divisor = c("r", "d"),ratio = FALSE)
}
\arguments{
\item{ppp.marks}{A "\verb{ppp}" object with two qualitative marks. One is a multiple category mark (for example, species in a forest) and the other is a dichotomous mark(for example, "Yes" and "No").}

\item{V.species}{A column in data represents a multiple category mark, such as tree species in a forest.}

\item{host}{A value of V.species, indicating the focal tree species. That is the host tree of a disease which is concerned.}

\item{V.bistate}{A column in data represents the bi-state of the host tree, for example, infected or uninfected.}

\item{yes}{The value of successful category in the mark \verb{V.bistate}, for example, "Yes". If \verb{missing}, the second level of the mark will be assigned.}

\item{no}{The value of failure category n the mark \verb{V.bistate}, for example, "No". If \verb{missing}, the first level of the mark will be assigned.}

\item{...}{Same as in \verb{pcfdot()} from the package \verb{spatstat}}

\item{r}{Same as in \verb{pcfdot()}from the package \verb{spatstat}}

\item{kernel}{Same as in \verb{pcfdot()} from the package \verb{spatstat}}

\item{bw}{Same as in \verb{pcfdot()} from the package \verb{spatstat}}

\item{stoyan}{Same as in \verb{pcfdot()} from the package \verb{spatstat}}

\item{correction}{Same as in \verb{pcfdot()} from the package \verb{spatstat}}

\item{divisor}{Same as in \verb{pcfdot()} from the package \verb{spatstat}}

\item{ratio}{Same as in \verb{pcfdot()} from the package \verb{spatstat}}
}
\value{
A object of "\verb{fv}".
}
\description{
Estimates difference between two dot-type pair correlation functions (\verb{pcfdot}).
Same as g_gdot.
}
\details{
Bivariate difference pcf (or \verb{g} function) is estimated as (Raventós, et al., 2010).
\deqn{g_{l,l+m}(r)-g_{m,l+m}(r)}
where \eqn{g()} is the  pair-correlation function, \emph{l} and \emph{m} are
two levels of a binary mark.

Use the function \code{\link{perm.r1fn.mkdf}} to permute \verb{V.bistate} mark of the
focal species \verb{host} for envelope simulation.
}
\examples{
\dontrun{
plot(g_gdot2(ppp.marks.c3),ylim = c(-3,0))
nsim=2499
ev.g_gdot2.ash.c3<-envelope(ppp.marks.c3,g_gdot2,nsim=nsim,
    funargs=list(V.species="Species", host="Ash", V.bistate="Infected"),
    simulate=expression(perm.r1fn.mkdf(ppp.marks.c3,V.species="Species",
    host="Ash",V.bistate="Infected")))
plot(ev.g_gdot2.ash.c3,ylim=c(-4,4))
}
}
\references{
Wiegand, T. and Moloney, K. A. 2013. Handbook of spatial point-pattern
analysis in ecology. Chapman and Hall/CRC, Boca Raton.

Raventós, J., Wiegand, T., and de Luis, M. 2010. Evidence for the spatial
segregation hypothesis: a test with nine-year survivorship data in a
Mediterranean shrubland. Ecology. 91 7:2110-2120.
}
\seealso{
\code{\link{g_gdot}} which does the same job,
\code{\link{K_Kdot2}}, \code{\link{L_Ldot2}}, \code{\link{J_Jdot2}}, \code{\link{ND_NDdot2}},
\code{\link{trig_g}}, \code{\link{trigcross}}.
}
\author{
Tianzhong Jing \email{Jingtianzhong@163.com}
}
