% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/govergdot.R
\name{govergdot}
\alias{govergdot}
\title{\verb{pcfdot} ratio estimation, tidy data}
\usage{
govergdot(ppp.marks,yes="Yes",no="No",...,
    r = NULL,
    kernel = "epanechnikov", bw = NULL, stoyan = 0.15,
    correction = c("isotropic", "Ripley", "translate"),
    divisor = c("r", "d"),ratio = FALSE)
}
\arguments{
\item{ppp.marks}{A "\verb{ppp}" object with a dichotomous mark(for example, "Yes" and "No").}

\item{yes}{The value of successful category in the dichotomous mark, for example, "Yes".
If \verb{missing}, the second level of the mark will be assigned.}

\item{no}{The value of failure category in the dichotomous mark, for example, "No".
If \verb{missing}, the first level of the mark will be assigned.}

\item{...}{Same as in \verb{pcfdot()} from the package \verb{spatstat}}

\item{r}{Same as in \verb{pcfdot()} from the package \verb{spatstat}}

\item{kernel}{Same as in \verb{pcfdot()} from the package \verb{spatstat}}

\item{bw}{Same as in \verb{pcfdot()} from the package \verb{spatstat}}

\item{stoyan}{Same as in \verb{pcfdot()} from the package \verb{spatstat}}

\item{correction}{Same as in \verb{pcfdot()} from the package \verb{spatstat}}

\item{divisor}{Same as in \verb{pcfdot()} from the package \verb{spatstat}}

\item{ratio}{Same as in \verb{pcfdot()} from the package \verb{spatstat}}
}
\value{
A object of "\verb{fv}".
}
\description{
Estimates the ratio of two dot-type pair correlation functions (\verb{pcfdot}).
}
\details{
Ratio of two pcfdot fuctions(or \emph{g} function) is estimated as (Wiegand, 2018).
\deqn{\frac{g_{l,l+m}(r)}{g_{m,l+m}(r)}}
where \eqn{g()} is the  pair-correlation function, \emph{l} and \emph{m} are
two levels of a binary mark.

Estimates strength of density on mortality.

Use the function \verb{rlabel()} from the package \verb{spatstat} to shuffle
the binary mark for envelope simulation.
}
\examples{
\dontrun{
plot(govergdot(ppp.ash.c3))
nsim=19
ev.govergdot.ash.c3<-envelope(ppp.ash.c3,govergdot,nsim=nsim,
    funargs=list(yes="Yes",no="No"),simulate=rlabel)
plot(ev.govergdot.ash.c3)
}

}
\references{
Wiegand, T. and Moloney, K. A. 2013. Handbook of spatial point-pattern
analysis in ecology. Chapman and Hall/CRC, Boca Raton.

Wiegand, T. 2018. User Manual for the Programita software.
}
\seealso{
\code{\link{govergdot2}} which does the same job,
\code{\link{KoverKdot}}, \code{\link{LoverLdot}}, \code{\link{JoverJdot}}, \code{\link{NDoverNDdot}}
\code{\link{trigcross}}, \code{\link{trig_g}}.
}
\author{
Tianzhong Jing \email{Jingtianzhong@163.com}
}
