% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/J_Jcross2.R
\name{J_Jcross2}
\alias{J_Jcross2}
\title{\verb{Jcross} difference estimation, untidy data}
\usage{
J_Jcross2(ppp.marks,V.species="Species", host="Ash", V.bistate="Infected",
    yes="Yes",no="No",t.type="yn-nn",...,
    eps=NULL, r=NULL, breaks=NULL, correction=NULL)
}
\arguments{
\item{ppp.marks}{A "\verb{ppp}" object with at least two qualitative marks. One is a multiple category mark (for example, species in a forest) and the other is a dichotomous mark (for example, "Yes" and "No").}

\item{V.species}{A column in data represents a multiple category mark, such as tree species in a forest.}

\item{host}{A value of V.species, indicating the focal tree species. That is the host tree of a disease which is concerned.}

\item{V.bistate}{A column in data represents the bi-state of the host tree, for example, infected or uninfected.}

\item{yes}{The value of successful category in the mark \verb{V.bistate}, for example, "Yes". If \verb{missing}, the second level of the mark will be assigned.}

\item{no}{The value of failure category n the mark \verb{V.bistate}, for example, "No". If \verb{missing}, the first level of the mark will be assigned.}

\item{t.type}{Character, indicating what type of test to be estimated. The value can be "yn-nn", "ny-nn", "yy-nn", and "ny-yn",see \verb{Details}}

\item{...}{Same as in \verb{Jcross()} from the package \verb{spatstat}}

\item{eps}{Same as in \verb{Jcross()} from the package \verb{spatstat}}

\item{r}{Same as in \verb{Jcross()} from the package \verb{spatstat}}

\item{breaks}{Same as in \verb{Jcross()} from the package \verb{spatstat}}

\item{correction}{Same as in \verb{Jcross()} from the package \verb{spatstat}}
}
\value{
A object of "\verb{fv}".
}
\description{
Estimates difference between two cross-type \verb{J} functions (\verb{Jcross}).
}
\details{
Bivariate difference \emph{J} function is estimated as (Raventós, et al., 2010).

if t.type=="yn-nn", \deqn{J_{m,l}(r)-J_{l,l}(r)}
if t.type=="ny-nn", \deqn{J_{l,m}(r)-J_{l,l}(r)}
if t.type=="yy-nn", \deqn{J_{m,m}(r)-J_{l,l}(r)}
if t.type=="ny-yn", \deqn{J_{l,l}(r)-J_{m,l}(r)}
where \eqn{J()} is the \verb{J} function, \emph{l} and \emph{m} are
two levels of a binary mark.

Performs the same job as \code{\link{J_Jcross}}, but for untidy data,
i.e., the mark is a data frame. The mark variable (for example, \verb{Species})
contains a focal tree species and other tree species.
Another mark variable \verb{V.bistate} is only connected to the focal species.

This function tests, for example, the host density-dependent infection of a
forest disease.

Use the function \code{\link{perm.r1fn.mkdf}} to permute \verb{V.bistate} mark of the
focal species for envelope simulation.
}
\examples{
\dontrun{
plot(J_Jcross2(ppp.marks.c3,t.type="ny-yn"))
nsim=2499
ev.j_jcross2.ash.c3<-envelope(ppp.marks.c3,J_Jcross2,nsim=nsim,
    funargs=list(V.species="Species", host="Ash", V.bistate="Infected"),
    simulate=expression(perm.r1fn.mkdf(ppp.marks.c3,V.species="Species",
    host="Ash",V.bistate="Infected")))
plot(ev.j_jcross2.ash.c3)
}

}
\references{
Wiegand, T. and Moloney, K. A. 2013. Handbook of spatial point-pattern
analysis in ecology. Chapman and Hall/CRC, Boca Raton.

Raventós, J., Wiegand, T., and de Luis, M. 2010. Evidence for the spatial
segregation hypothesis: a test with nine-year survivorship data in a
Mediterranean shrubland. Ecology. 91 7:2110-2120.
}
\seealso{
\code{\link{J_Jcross}} which does the same job,
\code{\link{K_Kcross2}}, \code{\link{L_Lcross2}}, \code{\link{g_gcross2}},\code{\link{ND_NDcross2}},
\code{\link{triJcross}}, \code{\link{triJ_J}}.
}
\author{
Tianzhong Jing \email{Jingtianzhong@163.com}
}
