% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/J_Jdot.R
\name{J_Jdot}
\alias{J_Jdot}
\title{\verb{Jdot} difference estimation, tidy data}
\usage{
J_Jdot(ppp.marks,yes="Yes",no="No",...,
    eps=NULL, r=NULL, breaks=NULL, correction=NULL)
}
\arguments{
\item{ppp.marks}{A "\verb{ppp}" object with a dichotomous mark(for example, "Yes" and "No").}

\item{yes}{The value of successful category in the dichotomous mark, for example, "Yes".
If \verb{missing}, the second level of the mark will be assigned.}

\item{no}{The value of failure category in the dichotomous mark, for example, "No".
If \verb{missing}, the first level of the mark will be assigned.}

\item{...}{Same as in \verb{Jdot()} from the package \verb{spatstat}}

\item{eps}{Same as in \verb{Jdot()} from the package \verb{spatstat}}

\item{r}{Same as in \verb{Jdot()} from the package \verb{spatstat}}

\item{breaks}{Same as in \verb{Jdot()} from the package \verb{spatstat}}

\item{correction}{Same as in \verb{Jdot()} from the package \verb{spatstat}}
}
\value{
A object of "\verb{fv}".
}
\description{
Estimates difference between two dot-type \emph{J} functions (\verb{Jdot}).
}
\details{
Bivariate difference \emph{J} functions is estimated as (Raventós, et al., 2010).
\deqn{J_{l,l+m}(r)-J_{m,l+m}(r)}
where \eqn{J()} is \emph{J} function, \emph{l} and \emph{m} are
two levels of a binary mark.

This function tests, for example, the host density-dependent infection of a forest disease.


Use the function \verb{rlabel} from the package \verb{spatstat} to shuffle
the binary mark for envelope simulation.
}
\examples{
\dontrun{
plot(J_Jdot(ppp.ash.c3))
nsim=19
ev.j_jdot.ash.c3<-envelope(ppp.ash.c3,J_Jdot,nsim=nsim,
    funargs=list(yes="Yes",no="No"),simulate=rlabel)
plot(ev.j_jdot.ash.c3)
}
}
\references{
Wiegand, T. and Moloney, K. A. 2013. Handbook of spatial point-pattern
analysis in ecology. Chapman and Hall/CRC, Boca Raton.

Raventós, J., Wiegand, T., and de Luis, M. 2010. Evidence for the spatial
segregation hypothesis: a test with nine-year survivorship data in a
Mediterranean shrubland. Ecology. 91 7:2110-2120.
}
\seealso{
\code{\link{J_Jdot2}} which does the same job,
\code{\link{K_Kdot}}, \code{\link{L_Ldot}}, \code{\link{g_gdot}},\code{\link{ND_NDdot}},
\code{\link{triJcross}}, \code{\link{triJ_J}}.
}
\author{
Tianzhong Jing \email{Jingtianzhong@163.com}
}
