% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JoverJdot.R
\name{JoverJdot}
\alias{JoverJdot}
\title{\verb{Jdot} ratio estimation, tidy data}
\usage{
JoverJdot(ppp.marks,yes="Yes",no="No",...,
    eps=NULL, r=NULL, breaks=NULL, correction=NULL)
}
\arguments{
\item{ppp.marks}{A "\verb{ppp}" object with a dichotomous mark(for example, "Yes" and "No").}

\item{yes}{The value of successful category in the dichotomous mark, for example, "Yes".
If \verb{missing}, the second level of the mark will be assigned.}

\item{no}{The value of failure category in the dichotomous mark, for example, "No".
If \verb{missing}, the first level of the mark will be assigned.}

\item{...}{Same as in \verb{Jdot()} from the package \verb{spatstat}}

\item{eps}{Same as in \verb{Jdot()} from the package \verb{spatstat}}

\item{r}{Same as in \verb{Jdot()} from the package \verb{spatstat}}

\item{breaks}{Same as in \verb{Jdot()} from the package \verb{spatstat}}

\item{correction}{Same as in \verb{Jdot()} from the package \verb{spatstat}}
}
\value{
A object of "\verb{fv}".
}
\description{
Estimates the ratio two dot-type \emph{J} functions (\verb{Jdot}).
}
\details{
Ratio of two \verb{Jdot} fuctions is estimated as (Wiegand, 2018).
\deqn{\frac{J_{l,l+m}(r)}{J_{m,l+m}(r)}}
where \eqn{J()} is the \verb{J} function, \emph{l} and \emph{m} are
two levels of a binary mark.

This function tests the strength of density on mortality.

Use the function \verb{rlabel} from the package \verb{spatstat} to shuffle
the binary mark for envelope simulation.
}
\examples{
\dontrun{
plot(JoverJdot(ppp.ash.c3))
nsim=2499
ev.joverjdot.ash.c3<-envelope(ppp.ash.c3,JoverJdot,nsim=nsim,
    funargs=list(yes="Yes",no="No"),simulate=rlabel)
plot(ev.joverjdot.ash.c3)
}

}
\references{
Wiegand, T. and Moloney, K. A. 2013. Handbook of spatial point-pattern
analysis in ecology. Chapman and Hall/CRC, Boca Raton.

Wiegand, T. 2018. User Manual for the Programita software.
}
\seealso{
\code{\link{JoverJdot2}} which does the same job,
\code{\link{govergdot}}, \code{\link{KoverKdot}}, \code{\link{LoverLdot}},
\code{\link{NDoverNDdot}}.
}
\author{
Tianzhong Jing \email{Jingtianzhong@163.com}
}
