% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JoverJdot2.R
\name{JoverJdot2}
\alias{JoverJdot2}
\title{\verb{Jdot} ratio estimation, untidy data}
\usage{
JoverJdot2(ppp.marks,V.species="Species", host="Ash",
    V.bistate="Infected",yes="Yes",no="No",...,
    eps=NULL, r=NULL, breaks=NULL, correction=NULL)
}
\arguments{
\item{ppp.marks}{A "\verb{ppp}" object with two qualitative marks. One is a multiple category mark (for example, species in a forest) and the other is a dichotomous mark(for example, "Yes" and "No").}

\item{V.species}{A column in data represents a multiple category mark, such as tree species in a forest.}

\item{host}{A value of V.species, indicating the focal tree species. That is the host tree of a disease which is concerned.}

\item{V.bistate}{A column in data represents the bi-state of the host tree, for example, infected or uninfected.}

\item{yes}{The value of successful category in the mark \verb{V.bistate}, for example, "Yes". If \verb{missing}, the second level of the mark will be assigned.}

\item{no}{The value of failure category n the mark \verb{V.bistate}, for example, "No". If \verb{missing}, the first level of the mark will be assigned.}

\item{...}{Same as in \verb{Jdot()} from the package \verb{spatstat}}

\item{eps}{Same as in \verb{Jdot()} from the package \verb{spatstat}}

\item{r}{Same as in \verb{Jdot()} from the package \verb{spatstat}}

\item{breaks}{Same as in \verb{Jdot()} from the package \verb{spatstat}}

\item{correction}{Same as in \verb{Jdot()} from the package \verb{spatstat}}
}
\value{
A object of "\verb{fv}".
}
\description{
Estimates the ratio of two dot-type \emph{J} functions (\verb{Jdot}).
}
\details{
Ratio of two \verb{Jdot} fuctions is estimated as (Wiegand, 2018).
\deqn{\frac{J_{l,l+m}(r)}{J_{m,l+m}(r)}}
where \eqn{J()} is the \emph{J} function, \emph{l} and \emph{m} are
two levels of a binary mark.

This function tests the strength of density on mortality.

Use the function \code{\link{perm.r1fn.mkdf}} to permute \verb{V.bistate} mark of the
focal species \verb{host} for envelope simulation.
}
\examples{
\dontrun{
plot(JoverJdot2(ppp.marks.c3))
nsim=19
ev.jovjdot2.ash.c3<-envelope(ppp.marks.c3,JoverJdot2,nsim=nsim,
    funargs=list(V.species="Species", host="Ash", V.bistate="Infected"),
    simulate=expression(perm.r1fn.mkdf(ppp.marks.c3,V.species="Species",
    host="Ash",V.bistate="Infected")))
plot(ev.jovjdot2.ash.c3)
}

}
\references{
Wiegand, T. and Moloney, K. A. 2013. Handbook of spatial point-pattern
analysis in ecology. Chapman and Hall/CRC, Boca Raton.

Wiegand, T. 2018. User Manual for the Programita software.
}
\seealso{
\code{\link{JoverJdot}} which does the same job,
\code{\link{govergdot2}}, \code{\link{KoverKdot2}}, \code{\link{LoverLdot2}},
\code{\link{NDoverNDdot2}}.
}
\author{
Tianzhong Jing \email{Jingtianzhong@163.com}
}
