% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KoverKdot.R
\name{KoverKdot}
\alias{KoverKdot}
\title{\verb{Kdot} ratio estimation, tidy data}
\usage{
KoverKdot(ppp.marks,yes="Yes",no="No",...,
    r=NULL, breaks=NULL, correction,
    ratio=FALSE)
}
\arguments{
\item{ppp.marks}{A "\verb{ppp}" object with a dichotomous mark(for example, "Yes" and "No").}

\item{yes}{The value of successful category in the dichotomous mark, for example, "Yes".
If \verb{missing}, the second level of the mark will be assigned.}

\item{no}{The value of failure category in the dichotomous mark, for example, "No".
If \verb{missing}, the first level of the mark will be assigned.}

\item{...}{Same as in \verb{Kdot()} from the package \verb{spatstat}}

\item{r}{Same as in \verb{Kdot()} from the package \verb{spatstat}}

\item{breaks}{Same as in \verb{Kdot()} from the package \verb{spatstat}}

\item{correction}{Same as in \verb{Kdot()} from the package \verb{spatstat}}

\item{ratio}{Same as in \verb{Kdot()} from the package \verb{spatstat}}
}
\value{
A object of "\verb{fv}".
}
\description{
Estimates the ratio two dot-type \emph{K} functions (\verb{Kdot}).
}
\details{
Ratio of two \verb{Kdot} fuctions is estimated as (Wiegand, 2018).
\deqn{\frac{K_{l,l+m}(r)}{K_{m,l+m}(r)}}
where \eqn{K()} is the Ripley \emph{K} function, \emph{l} and \emph{m} are
two levels of a binary mark.

This function tests the strength of density on mortality.

Use the function \verb{rlabel()} from the package \verb{spatstat} to shuffle
the binary mark for envelope simulation.
}
\examples{
\dontrun{
plot(KoverKdot(ppp.ash.c3))
nsim=2499
ev.koverkdot.ash.c3<-envelope(ppp.ash.c3,KoverKdot,nsim=nsim,
    funargs=list(yes="Yes",no="No"),simulate=rlabel)
plot(ev.koverkdot.ash.c3)
}

}
\references{
Wiegand, T. and Moloney, K. A. 2013. Handbook of spatial point-pattern
analysis in ecology. Chapman and Hall/CRC, Boca Raton.

Wiegand, T. 2018. User Manual for the Programita software.
}
\seealso{
\code{\link{KoverKdot2}} which does the same job,
\code{\link{govergdot}}, \code{\link{LoverLdot}}, \code{\link{JoverJdot}}, \code{\link{NDoverNDdot}}
\code{\link{triK_K}}, \code{\link{triKcross}}.
}
\author{
Tianzhong Jing \email{Jingtianzhong@163.com}
}
