% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_Lcross.R
\name{L_Lcross}
\alias{L_Lcross}
\title{\verb{Lcross} difference estimation, tidy data}
\usage{
L_Lcross(ppp.marks,yes="Yes",no="No",t.type="yn-nn",...,
    r=NULL, breaks=NULL, correction,
    ratio=FALSE)
}
\arguments{
\item{ppp.marks}{A "\verb{ppp}" object with a dichotomous mark (for example, "Yes" and "No").}

\item{yes}{The value of successful category in the dichotomous mark, for example, "Yes". If \verb{missing}, the second level of the mark will be assigned.}

\item{no}{The value of failure category in the dichotomous mark, for example, "No". If \verb{missing}, the first level of the mark will be assigned.}

\item{t.type}{Character, indicating what type of test to be estimated. The value can be "yn-nn", "ny-nn", "yy-nn", and "ny-yn",see \verb{Details}}

\item{...}{Same as in \verb{Kcross()} from the package \verb{spatstat}}

\item{r}{Same as in \verb{Kcross()} from the package \verb{spatstat}}

\item{breaks}{Same as in \verb{Kcross()} from the package \verb{spatstat}}

\item{correction}{Same as in \verb{Lcross()} from the package \verb{spatstat}}

\item{ratio}{Same as in \verb{Kcross()} from the package \verb{spatstat}}
}
\value{
A object of "\verb{fv}".
}
\description{
Estimates difference between two cross-type \emph{L} functions (\verb{Lcross}).
}
\details{
Bivariate difference \emph{L} functions is estimated as (Raventós, et al., 2010).

if t.type=="yn-nn", \deqn{L_{m,l}(r)-L_{l,l}(r)}
if t.type=="ny-nn", \deqn{L_{l,m}(r)-L_{l,l}(r)}
if t.type=="yy-nn", \deqn{L_{m,m}(r)-L_{l,l}(r)}
if t.type=="ny-yn", \deqn{L_{l,l}(r)-L_{m,l}(r)}
where \eqn{L()} is the \emph{L}  function, \emph{l} and \emph{m} are
two levels of a binary mark.

This function tests, for example, the host density-dependent infection of a
forest disease.

Performs the same job as \code{\link{L_Lcross2}}, but used for tidy data,
i.e, with only one mark.

Use the function \verb{rlabel()} from the package \verb{spatstat} to permute
the binary mark for envelope simulation.
}
\examples{
\dontrun{
plot(L_Lcross(ppp.ash.c3,t.type="yy-nn"))
nsim=2499
ev.l_lcross.ash.c3<-envelope(ppp.ash.c3,L_Lcross,nsim=nsim,
    funargs=list(yes="Yes",no="No"),simulate=rlabel)
plot(ev.l_lcross.ash.c3)
}

}
\references{
Wiegand, T. and Moloney, K. A. 2013. Handbook of spatial point-pattern
analysis in ecology. Chapman and Hall/CRC, Boca Raton.

Raventós, J., Wiegand, T., and de Luis, M. 2010. Evidence for the spatial
segregation hypothesis: a test with nine-year survivorship data in a
Mediterranean shrubland. Ecology. 91 7:2110-2120.
}
\seealso{
\code{\link{L_Lcross2}} which does the same job,
\code{\link{g_gcross}}, \code{\link{K_Kcross}}, \code{\link{J_Jcross}},\code{\link{ND_NDcross}},
\code{\link{triKcross}}, \code{\link{triL_L}}.
}
\author{
Tianzhong Jing \email{Jingtianzhong@163.com}
}
