% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_Ldot2.R
\name{L_Ldot2}
\alias{L_Ldot2}
\title{\verb{Ldot} difference estimation, untidy data}
\usage{
L_Ldot2(ppp.marks,V.species="Species", host="Ash", V.bistate="Infected",
    yes="Yes",no="No",...,
    r=NULL, breaks=NULL, correction,
    ratio=FALSE)
}
\arguments{
\item{ppp.marks}{A "\verb{ppp}" object with two qualitative marks. One is a multiple category mark (for example, species in a forest) and the other is a dichotomous mark(for example, "Yes" and "No").}

\item{V.species}{A column in data represents a multiple category mark, such as tree species in a forest.}

\item{host}{A value of V.species, indicating the focal tree species. That is the host tree of a disease which is concerned.}

\item{V.bistate}{A column in data represents the bi-state of the host tree, for example, infected or uninfected.}

\item{yes}{The value of successful category in the mark \verb{V.bistate}, for example, "Yes". If \verb{missing}, the second level of the mark will be assigned.}

\item{no}{The value of failure category n the mark \verb{V.bistate}, for example, "No". If \verb{missing}, the first level of the mark will be assigned.}

\item{...}{Same as in \verb{Kdot()} from the package \verb{spatstat}}

\item{r}{Same as in \verb{Kdot()} from the package \verb{spatstat}}

\item{breaks}{Same as in \verb{Kdot()} from the package \verb{spatstat}}

\item{correction}{Same as in \verb{Ldot()} from the package \verb{spatstat}}

\item{ratio}{Same as in \verb{Kdot()} from the package \verb{spatstat}}
}
\value{
A object of "\verb{fv}".
}
\description{
Estimates difference between two dot-type \emph{L} functions (\verb{Ldot}).
}
\details{
Bivariate difference \emph{L} function is estimated as (Raventós, et al., 2010).
\deqn{L_{l,l+m}(r)-L_{m,l+m}(r)}
where \eqn{L()} is the \emph{L} function, \emph{l} and \emph{m} are
two levels of a binary mark.

This function tests, for example, host density-dependent infection of a forest disease.

Use the function \code{\link{perm.r1fn.mkdf}} to permute \verb{V.bistate} mark of the
focal species \verb{host} for envelope simulation.
}
\examples{
#plot(L_Ldot2(ppp.marks.c3))
#nsim=2499
#ev.l_ldot2.ash.c3<-envelope(ppp.marks.c3,L_Ldot2,nsim=nsim,
#    funargs=list(V.species="Species", host="Ash", V.bistate="Infected"),
#    simulate=expression(perm.r1fn.mkdf(ppp.marks.c3,V.species="Species",
#    host="Ash",V.bistate="Infected")))
#plot(ev.l_ldot2.ash.c3)

}
\references{
Wiegand, T. and Moloney, K. A. 2013. Handbook of spatial point-pattern
analysis in ecology. Chapman and Hall/CRC, Boca Raton.

Raventós, J., Wiegand, T., and de Luis, M. 2010. Evidence for the spatial
segregation hypothesis: a test with nine-year survivorship data in a
Mediterranean shrubland. Ecology. 91 7:2110-2120.
}
\seealso{
\code{\link{L_Ldot}} which does the same job,\code{\link{K_Kdot}}
\code{\link{g_gdot2}}, \code{\link{K_Kdot2}}, \code{\link{J_Jdot2}}, \code{\link{ND_NDdot2}}
\code{\link{triL_L}}, \code{\link{triLcross}}.
}
\author{
Tianzhong Jing \email{Jingtianzhong@163.com}
}
