% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkcorr_d7.R
\name{mkcorr.d7}
\alias{mkcorr.d7}
\title{\verb{markcorr} for data 7 tests 1-5}
\usage{
mkcorr.d7(X, m.qti.1="DBH", m.qti.2="Disease",
    f = function(m1, m2) {m1*m2}, r=NULL,
    correction=c("isotropic", "Ripley", "translate"),
    method="density", ...,
    weights=NULL, f1=NULL, normalise=TRUE, fargs=NULL, internal=NULL)
}
\arguments{
\item{X}{An object of class "\verb{ppp}"，with two quantitative marks.}

\item{m.qti.1}{A mark in data "\verb{X}" indicating the first quantitative mark.}

\item{m.qti.2}{A mark in data "\verb{X}" indicating the second quantitative mark.}

\item{f}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{r}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{correction}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{method}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{...}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{weights}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{f1}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{normalise}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{fargs}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{internal}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}
}
\value{
A object of "\verb{fv}".
}
\description{
Estimates mark correlation functions for univariate pattern with
two quantitative marks. The test functions include mark correlation function,
r-mark correlation function,mark variogram, and Moran I.
}
\details{
The data type and test function are referred to (Wiegand and Moloney, 2013).
Data 7 is an univariate pattern with two quantitative marks.

The following test functions (tests 1-5) have been tested.
\itemize{
    \item f = function(m1,m2) {m1*m2}
    \item f = function(m1,m2) {m1}
    \item f = function(m1,m2) {m2}
    \item f = function(m1,m2) {0.5*(m1-m2)^2}
    \item f = function(m1,m2) {(m1-mean(m1))*(m2-mean(m2))}
}
Other f(m1,m2) functions may also work.

The function \verb{rlabel()} from the package \verb{spatstat} may be used to
permute the marks for envelope simulation.
}
\examples{
\dontrun{
nsim=2499
ev.mkc.m1m2.insectnum.dbh.g1<-envelope(ppp.ash.g1,mkcorr.d7,
    funargs=list(m.qti.1="dbh",m.qti.2="count_insect",
    f = function(m1, m2) {m1*m2}), nsim=nsim,simulate=rlabel)
plot(ev.mkc.m1m2.insectnum.dbh.g1)

ev.mkc.vario.insectnum.dbh.g1<-envelope(ppp.ash.g1,mkcorr.d7,
    funargs=list(m.qti.1="dbh",m.qti.2="count_insect",
    f = function(m1, m2) {0.5*(m1-m2)^2}), nsim=nsim,simulate=rlabel)
plot(ev.mkc.vario.insectnum.dbh.g1)

ev.mkc.moran.insectnum.dbh.g1<-envelope(ppp.ash.g1,mkcorr.d7,
    funargs=list(m.qti.1="dbh",m.qti.2="count_insect",
    f = function(m1, m2) {(m1-mean(m1))*(m2-mean(m2))}), nsim=nsim,
    simulate=rlabel)
plot(ev.mkc.moran.insectnum.dbh.g1)
}

}
\references{
Wiegand, T. and Moloney, K. A. 2013. Handbook of spatial point-pattern
analysis in ecology. Chapman and Hall/CRC, Boca Raton.

Wiegand, T. 2018. User Manual for the Programita software.
}
\seealso{
\code{\link{mkcorr.d7.t6}}, \code{\link{mkcorr.d8}},
\code{\link{mkcorr.d9}}, \code{\link{mkcorr.t7}} ,\code{\link{mkcorr.t8}}
}
\author{
Tianzhong Jing \email{Jingtianzhong@163.com}
}
