% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkcorr_d7_t6.R
\name{mkcorr.d7.t6}
\alias{mkcorr.d7.t6}
\title{\verb{markcorr} for data 7 test 6}
\usage{
mkcorr.d7.t6(X, m.qti.1="DBH", m.qti.2="Disease", r=NULL,
    correction=c("isotropic", "Ripley", "translate"),
    method="density", ...,
    weights=NULL, normalise=TRUE, internal=NULL)
}
\arguments{
\item{X}{An object of class "\verb{ppp}"，with two quantitative marks.}

\item{m.qti.1}{A mark of data "\verb{X}" indicating the first quantitative mark.}

\item{m.qti.2}{A mark of data "\verb{X}" indicating the second quantitative mark.}

\item{r}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{correction}{Same as \verb{markcorr()} from the package \verb{spatstat}.}

\item{method}{Same as \verb{markcorr()} from the package \verb{spatstat}.}

\item{...}{Same as \verb{markcorr()} from the package \verb{spatstat}.}

\item{weights}{Same as \verb{markcorr()} from the package \verb{spatstat}.}

\item{normalise}{Same as \verb{markcorr()} from the package \verb{spatstat}.}

\item{internal}{Same as \verb{markcorr()} from the package \verb{spatstat}.}
}
\value{
A object of "\verb{fv}".
}
\description{
Estimates Schlather I mark correlation function for univariate pattern
with two quantitative marks.
}
\details{
The data type and test function are referred to (Wiegand and Moloney, 2013).
Data 7 is an univariate pattern with two quantitative marks. Test 6 is Schlather I:
\deqn{t(m_{il},m_{jm})=(m_{il}-\mu_l(r))(m_{jm}-\mu_m(r))}

This estimation is time-cost.

The function \verb{rlabel()} from the package \verb{spatstat} may be used to
permute the marks for envelope simulation.
}
\examples{
\dontrun{
plot(mkcorr.d7.t6(ppp.ash.g1,m.qti.1="dbh",m.qti.2="count_insect"))
nsim=2499
ev.mkc.t6.insectnum.dbh.g1<-envelope(ppp.ash.g1,mkcorr.d7.t6,
    funargs=list(m.qti.1="dbh",m.qti.2="count_insect"), nsim=nsim,
    simulate=rlabel)
plot(ev.mkc.t6.insectnum.dbh.g1)
}

}
\references{
Wiegand, T. and Moloney, K. A. 2013. Handbook of spatial point-pattern
analysis in ecology. Chapman and Hall/CRC, Boca Raton.

Wiegand, T. 2018. User Manual for the Programita software.
}
\seealso{
\code{\link{mkcorr.d7}}, \code{\link{mkcorr.d8.t6}},
\code{\link{mkcorr.d9.t6}}, \code{\link{mkcorr.t7}} ,\code{\link{mkcorr.t8}}
}
\author{
Tianzhong Jing \email{Jingtianzhong@163.com}
}
