% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkcorr_d7_t8.R
\name{mkcorr.d7.t8}
\alias{mkcorr.d7.t8}
\title{Density Correlation Function, data 7}
\usage{
mkcorr.d7.t8(X, m.qti.1="DBH", r=NULL, delta=NULL, stoyan=0.15, method="density",
    weights=NULL,normalise=TRUE, internal=NULL,...)
}
\arguments{
\item{X}{An object of class "\verb{ppp}"，with one quantitative mark.}

\item{m.qti.1}{A mark in data "\verb{X}" indicating the quantitative mark for estimation.}

\item{r}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{delta}{Same as in \verb{localpcf()} from the package \verb{spatstat}.}

\item{stoyan}{Same as in \verb{localpcf()} from the package \verb{spatstat}.}

\item{method}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{weights}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{normalise}{Same as in \verb{markcorr()} from the package \verb{spatstat}}

\item{internal}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{...}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}
}
\value{
A object of "\verb{fv}".
}
\description{
Estimates density correlation function for univariate pattern with one quantitative marks.
}
\details{
The test function (test 8), is
\deqn{C_{m,g}(r)=(m_{il}-\mu_{l})(g_i(r)-g(r))}
where the function \eqn{g_i()} is the local pair-correlation function of
point \emph{i} and \eqn{g()} is the pair-correlation function which equals
to the mean of \eqn{g_i()}, \eqn{m_{il}} is the mark value of point \emph{i}
of mark \emph{l} and \eqn{\mu_{l}} is the mean mark \emph{l} of all points.

So this test is as same as \code{\link{mkcorr.t8}}. The only difference is that it
is applied to the data of more than one mark and can assign which mark to be tested.

This test deals with only one mark, no correlation between two marks.

This function is time consuming.
}
\examples{
\dontrun{
mkc.d7t8.g1<-mkcorr.d7.t8(ppp.ash.g1, m.qti.1="dbh")
plot(mkc.d7t8.g1)
nsim=2499
ev.mkc.d7t8.g1<-envelope(ppp.ash.g1, m.qti.1="dbh",mkcorr.d7.t8,
    nsim=nsim,
    simulate=rlabel)
plot(ev.mkc.d7t8.g1)
ev.mkc.d7t8.g1<-envelope(ppp.ash.g1, m.qti.1="dbh", mkcorr.d7.t8,
    nsim=nsim)
}

}
\references{
Fedriani, J, T. Wiegand, G. Calvo, A. Suarez-Esteban, M. Jacome, M. Zywiec,
and M.Delibes. 2015. Unravelling conflicting density- and distance-dependent
effects on plant reproduction using a spatially explicit approach.
Journal of Ecology: 103: 1344-1353

Wiegand, T. 2018. User Manual for the Programita software.
}
\seealso{
\code{\link{mkcorr.d7.t7}}, \code{\link{mkcorr.d8.t8}},
\code{\link{mkcorr.t8}}, \code{\link{mkcorr.d7.t6}}
}
\author{
Tianzhong Jing \email{Jingtianzhong@163.com}
}
