% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkcorr_d8.R
\name{mkcorr.d8}
\alias{mkcorr.d8}
\title{\verb{markcorr} for data 8 tests 1-5}
\usage{
mkcorr.d8(X, m.qli="Infected", m.qti="DBH",
    f = function(m1, m2) {m1*m2}, r=NULL,
    correction=c("isotropic", "Ripley", "translate"),
    method="density", ...,
    weights=NULL, f1=NULL, normalise=TRUE, fargs=NULL, internal=NULL)
}
\arguments{
\item{X}{An object of class "\verb{ppp}"，with one qualitative mark and one quantitative mark.}

\item{m.qli}{A mark of data "\verb{X}" indicates the qualitative mark.}

\item{m.qti}{A mark of data "\verb{X}" indicates the quantitative mark.}

\item{f}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{r}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{correction}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{method}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{...}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{weights}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{f1}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{normalise}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{fargs}{SSame as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{internal}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}
}
\value{
A object of "\verb{fv}".
}
\description{
Estimates mark correlation functions to an univariate pattern with
one quantitative mark and one qualitative mark. The test functions include
mark correlation function, r-mark correlation function,mark variogram,
and Moran's I.
}
\details{
Mark correlation functions for an univariate pattern with one quantitative mark and one qualitative mark


An example for this data structure is a pattern of surviving and dead trees
of a given species (qualitative mark), where the size of the trees is also
known (the quantitative mark). The basic interest in analyzing these types
of data is to find out whether the quantitative mark (e.g., size)
shows a spatial correlation with the qualitatively marked points
(i.e., surviving and dead)(Wiegand and Moloney,2013).

The data type and test function are referred to (Wiegand and Moloney, 2013).
Data 8 is an univariate pattern with one quantitative mark and one qualitative mark.

The following test functions (tests 1-5) have been tested.
\itemize{
    \item f = function(m1,m2) {m1*m2}
    \item f = function(m1,m2) {m1}
    \item f = function(m1,m2) {m2}
    \item f = function(m1,m2) {0.5*(m1-m2)^2}
    \item f = function(m1,m2) {(m1-mean(c(m1,m2)))*(m2-mean(c(m1,m2)))}
    \item f = function(m1,m2) {(m1-mean(m1))*(m2-mean(m2))}
}
The last test can be a second version of test 5.
Other f(m1,m2) functions may also work.

The function \code{\link{perm.onemkvar}} may be used to
permute one the two marks for envelope simulation.
}
\examples{
\dontrun{
nsim=2499
ev.mkc.m1m2.ash.infect.dbh.c3<-envelope(ppp.ash.infect.dbh.c3,mkcorr.d8,
    funargs=list(m.qli="Infected",m.qti="DBH",
    f = function(m1, m2) {m1 * m2}), nsim=nsim,
    simulate=expression(perm.onemkvar(ppp.ash.infect.dbh.c3,V.simul="DBH")))
plot(ev.mkc.m1m2.ash.infect.dbh.c3)

ev.mkc.MI.ash.infect.dbh.c3<-envelope(ppp.ash.infect.dbh.c3,mkcorr.d8,
    funargs=list(m.qli="Infected",m.qti="DBH",
    f = function(m1, m2) {(m1-mean(c(m1,m2)))*(m2-mean(c(m1,m2)))}), nsim=nsim,
    simulate=expression(perm.onemkvar(ppp.ash.infect.dbh.c3,V.simul="DBH")))
plot(ev.mkc.MI.ash.infect.dbh.c3)

ev.mkc.vario.ash.infect.dbh.c3<-envelope(ppp.ash.infect.dbh.c3,mkcorr.d8,
    funargs=list(m.qli="Infected",m.qti="DBH",
    f=function(m1,m2){0.5*(m1-m2)^2}), nsim=nsim,
    imulate=expression(perm.onemkvar(ppp.ash.infect.dbh.c3,V.simul="DBH")))
plot(ev.mkc.vario.ash.infect.dbh.c3)
}

}
\references{
Wiegand, T. and Moloney, K. A. 2013. Handbook of spatial point-pattern
analysis in ecology. Chapman and Hall/CRC, Boca Raton.

Raventós, J., Wiegand, T., and de Luis, M. 2010. Evidence for the spatial
segregation hypothesis: a test with nine-year survivorship data in a
Mediterranean shrubland. Ecology. 91 7:2110-2120.

Wiegand, T. 2018. User Manual for the Programita software.
}
\seealso{
\code{\link{mkcorr.d8.t6}}, \code{\link{mkcorr.d7}},
\code{\link{mkcorr.d9}}, \code{\link{mkcorr.t7}} ,\code{\link{mkcorr.t8}}
}
\author{
Tianzhong Jing \email{Jingtianzhong@163.com}
}
