% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkcorr_d8_t6.R
\name{mkcorr.d8.t6}
\alias{mkcorr.d8.t6}
\title{\verb{markcorr} for data 8 test 6}
\usage{
mkcorr.d8.t6(X, m.qli="Infected", m.qti="DBH",r=NULL,
    correction=c("isotropic", "Ripley", "translate"),
    method="density", ...,
    weights=NULL, normalise=TRUE, internal=NULL,mu.global=TRUE)
}
\arguments{
\item{X}{An object of class "\verb{ppp}"，with one qualitative mark and one quantitative mark.}

\item{m.qli}{A mark of data "\verb{X}" indicating the qualitative mark.}

\item{m.qti}{A mark of data "\verb{X}" indicating the quantitative mark.}

\item{r}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{correction}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{method}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{...}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{weights}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{normalise}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{internal}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{mu.global}{Logical, indicating whether mean is global or local.
If TRUE, \eqn{\mu=\mu_{l+m}} ; else \eqn{\mu=\mu_l} or\eqn{\mu=\mu_m}}
}
\value{
A object of "\verb{fv}".
}
\description{
Estimating Schlather I mark correlation function for an univariate pattern
with one quantitative mark and one qualitative mark.
}
\details{
The data type and test function are referred to (Wiegand and Moloney, 2013).
Data 8 is an univariate pattern with one quantitative mark and one qualitative mark.

This funciton estimates Schlather I mark correlation function (test 6) for data 8
(Wiegand and Moloney,2013). The function provides two versions of test 6.
If \verb{mu.global==TRUE},
\deqn{(m_{il}-\mu_{l+m}(r))(m_{jm}-\mu_{l+m}(r))}
else
\deqn{(m_{il}-\mu_l(r))(m_{jm}-\mu_m(r))}
where \emph{i} and \emph{j} represent points, and \emph{l} and \emph{m} represent marks.
And the result will be normalised by variance of the mark (1+2) and sd1*sd2,
respectively, if \verb{normalised==TRUE}.

The function \code{\link{perm.onemkvar}} may be used to
permute one the two marks for envelope simulation.
}
\examples{
\dontrun{
nsim=2499
ev.mkc.t6.ash.infect.dbh.c3<-envelope(ppp.ash.infect.dbh.c3,mkcorr.d8.t6,
    funargs=list(m.qli="Infected",m.qti="DBH"), nsim=nsim,
    simulate=expression(perm.onemkvar(ppp.ash.infect.dbh.c3,V.simul="DBH")))
plot(ev.mkc.t6.ash.infect.dbh.c3)
}

}
\references{
Wiegand, T. and Moloney, K. A. 2013. Handbook of spatial point-pattern
analysis in ecology. Chapman and Hall/CRC, Boca Raton.

Raventós, J., Wiegand, T., and de Luis, M. 2010. Evidence for the spatial
segregation hypothesis: a test with nine-year survivorship data in a
Mediterranean shrubland. Ecology. 91 7:2110-2120.

Wiegand, T. 2018. User Manual for the Programita software.
}
\seealso{
\code{\link{mkcorr.d8}}, \code{\link{mkcorr.d7.t6}},
\code{\link{mkcorr.d9.t6}}, \code{\link{mkcorr.t7}} ,\code{\link{mkcorr.t8}}
}
\author{
Tianzhong Jing \email{Jingtianzhong@163.com}
}
