% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkcorr_d8_t7.R
\name{mkcorr.d8.t7}
\alias{mkcorr.d8.t7}
\title{Cumulative Density Correlation Function, data 8}
\usage{
mkcorr.d8.t7(X, m.qli="Infected", m.qti="DBH", ind=TRUE,
    r=NULL, verbose = TRUE, correction="Ripley",
    method="density",
    weights=NULL,normalise=TRUE, internal=NULL, ...,
    mu.global=TRUE, dens.global=TRUE)
}
\arguments{
\item{X}{An object of class "\verb{ppp}"，with one quantitative mark.}

\item{m.qli}{A mark of data "\verb{X}" indicating the qualitative mark.}

\item{m.qti}{A mark of data "\verb{X}" indicating the quantitative mark.}

\item{ind}{logical, indicating the mark \verb{m.qti} of which level of \verb{m.qli} is used. If \verb{TRUE}, the mark corresponding to the second level of \verb{m.qli} is used;  else, the first level.}

\item{r}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{verbose}{Same as in \verb{localK()} from the package \verb{spatstat}.}

\item{correction}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{method}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{weights}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{normalise}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{internal}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{...}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{mu.global}{Logical, indicating whether mean is global or local.  If TRUE, \eqn{\mu=\mu_{l+m}} ; else \eqn{\mu=\mu_l} or\eqn{\mu=\mu_m}}

\item{dens.global}{Logical, indicating whether desnity is global or local.  If TRUE, \eqn{K(r)=K_{l+m}(r)} ; else \eqn{K(r)=K_l(r)} or\eqn{K(r)=K_m(r)}}
}
\value{
A object of "\verb{fv}".
}
\description{
Estimates cum. density correlation function for univariate pattern with one quantitative marks.
}
\details{
According to the variables \verb{mu.global} and \verb{dens.global}, there are
four types of the test function (test 7) for data 8:
\deqn{C_{m,K}(r)=(m_{il}-\mu_l)(K_{jm}(r)-K_m(r))}
\deqn{C_{m,K}(r)=(m_{il}-\mu_l)(K_{jm}(r)-K_{l+m}(r))}
\deqn{C_{m,K}(r)=(m_{il}-\mu_{l+m})(K_{jm}(r)-K_m(r))}
\deqn{C_{m,K}(r)=(m_{il}-\mu_{l+m})(K_{jm}(r)-K_{l+m}(r))}
where the function \eqn{K_i()} is the local Ripley \emph{K} function of
point \emph{i} and \eqn{K()} is the Ripley \emph{K} function which equals
to the mean of \eqn{K_i()}, \eqn{m_{il}} is the mark value of point \emph{i}
of mark \emph{l}, and \eqn{\mu_l} is mean of mark \emph{l} of all points.

This function is also suitable for data 9. In the case of data 9, both
\verb{mu.global} and \verb{dens.global} should be set to \verb{FALSE}.

This function is time consuming.
}
\examples{
\dontrun{
nsim=2499
ev.mkc.d8t7.c3<-envelope(ppp.ash.infect.dbh.c3, m.qli="Infected", m.qti="DBH",
    mkcorr.d8.t7,
    funargs=list(correction="translate",verbose =FALSE), nsim=nsim,
    simulate=NULL)
plot(ev.mkc.d8t7.c3)
}

}
\references{
Fedriani, J, T. Wiegand, G. Calvo, A. Suarez-Esteban, M. Jacome, M. Zywiec,
and M.Delibes. 2015. Unravelling conflicting density- and distance-dependent
effects on plant reproduction using a spatially explicit approach.
Journal of Ecology: 103: 1344-1353

Wiegand, T. 2018. User Manual for the Programita software.
}
\seealso{
\code{\link{mkcorr.d8.t8}}, \code{\link{mkcorr.d7.t7}},
\code{\link{mkcorr.d8.t6}}, \code{\link{mkcorr.t7}}.
}
\author{
Tianzhong Jing \email{Jingtianzhong@163.com}
}
