% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkcorr_d9.R
\name{mkcorr.d9}
\alias{mkcorr.d9}
\title{\verb{markcorr} for data 9 tests 1-5}
\usage{
mkcorr.d9(X, m.qli="Species", m.qti="Disease",
    f = function(m1, m2) {m1*m2}, r=NULL,
    correction=c("isotropic", "Ripley", "translate"),
    method="density", ...,
    weights=NULL, f1=NULL, normalise=TRUE,
    fargs=NULL, internal=NULL)
}
\arguments{
\item{X}{An object of class "\verb{ppp}", with one qualitative mark and one quantitative mark.}

\item{m.qli}{A mark of data "\verb{X}" indicating the qualitative mark.}

\item{m.qti}{A mark of data "\verb{X}" indicating the quantitative mark.}

\item{f}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{r}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{correction}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{method}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{...}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{weights}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{f1}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{normalise}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{fargs}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{internal}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}
}
\value{
A object of "\verb{fv}".
}
\description{
Estimates a series mark correlation function to a bivariate pattern with one quantitative mark. The test functions include mark correlation function, r-mark correlation function,mark variogram, and Moran's I.
}
\details{
The data type and test function are referred to (Wiegand and Moloney, 2013).
Data 9 is a bivariate pattern with one quantitative mark.

The following test functions (tests 1-5) have been tested.
\itemize{
    \item f = function(m1,m2) {m1*m2}
    \item f = function(m1,m2) {m1}
    \item f = function(m1,m2) {m2}
    \item f = function(m1,m2) {0.5*(m1-m2)^2}
    \item f = function(m1,m2) {(m1-mean(m1))*(m2-mean(m2))}#'
}
Other f(m1,m2) functions may also work.

In the case of data 9,a bivariate pattern with one quantitative mark,
the marks must only be randomized within each pattern (i.e., a point of
pattern 1 should not receive a mark of a point of pattern 2 and vice versa)
(Wiegand and Moloney, 2013). So employ the function \code{\link{perm.r1r2.mkdf}} to do
envelope simulation.
}
\examples{
#nsim=19
#ppp.marks.ash.birch.c4<-ppp.marks.sp4c.c4[[1]]
#ev.mkc.m1m2.ash.birch.disease.c4<-envelope(ppp.marks.ash.birch.c4,
#    mkcorr.d9,funargs=list(m.qli="Species",m.qti="Disease",
#    f = function(m1, m2) {m1 * m2}), nsim=nsim,
#    simulate=expression(perm.r1r2.mkdf(ppp.marks.ash.birch.c4,
 #   m.qli="Species",m.qti="Disease")))
#plot(ev.mkc.m1m2.ash.birch.disease.c4)

#ev.mkc.vario.ash.birch.disease.c4<-envelope(ppp.marks.ash.birch.c4,
#    mkcorr.d9,funargs=list(m.qli="Infected",m.qti="DBH",
#    f=function(m1,m2){0.5*(m1-m2)^2}), nsim=nsim,
#    simulate=expression(perm.r1r2.mkdf(ppp.marks.ash.birch.c4,
#    m.qli="Species",m.qti="Disease")))
#plot(ev.mkc.vario.ash.birch.disease.c4)

#ev.mkc.vario.c4<-anylapply(ppp.marks.sp4c.c4,envelope,mkcorr.d9,
#    funargs=list(m.qli="Species",m.qti="Disease",
 #   f=function(m1,m2){0.5*(m1-m2)^2}), nsim=nsim,
 #   simulate=perm.r1r2.mkdf)
#plot(ev.mkc.vario.c4, main="", main.panel=letters[1:6], legend=FALSE)

}
\references{
Wiegand, T. and Moloney, K. A. 2013. Handbook of spatial point-pattern
analysis in ecology. Chapman and Hall/CRC, Boca Raton.

Wiegand, T. 2018. User Manual for the Programita software.
}
\seealso{
\code{\link{mkcorr.d9.t6}}, \code{\link{mkcorr.d7}},
\code{\link{mkcorr.d8}}, \code{\link{mkcorr.t7}} ,\code{\link{mkcorr.t8}}
}
\author{
Tianzhong Jing \email{Jingtianzhong@163.com}
}
