% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkcorr_d9_t6.R
\name{mkcorr.d9.t6}
\alias{mkcorr.d9.t6}
\title{\verb{markcorr} for data 9 test 6}
\usage{
mkcorr.d9.t6(X, m.qli="Species", m.qti="Disease",r=NULL,
    correction=c("isotropic", "Ripley", "translate"),
    method="density", ...,
    weights=NULL, normalise=TRUE, internal=NULL)
}
\arguments{
\item{X}{An object of class "\verb{ppp}"，with one qualitative mark and one quantitative mark.}

\item{m.qli}{A mark of data "\verb{X}" indicating the qualitative mark.}

\item{m.qti}{A mark of data "\verb{X}" indicating the quantitative mark.}

\item{r}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{correction}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{method}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{...}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{weights}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{normalise}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{internal}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}
}
\value{
A object of "\verb{fv}".
}
\description{
Estimating Schlather I mark correlation function for a bivariate pattern
with one quantitative mark.
}
\details{
The data type and test function are referred to (Wiegand and Moloney, 2013).
Data 9 is a bivariate pattern with one quantitative mark.

This function tests (test 6)
\deqn{(m_{il}-\mu_l(r))(m_{jm}-\mu_m(r))}
The result will be normalised by sd1*sd2, if \verb{normalised==TRUE}.

In the case of data 9, a bivariate pattern with one quantitative mark,
the marks must only be randomized within each pattern (i.e., a point of
pattern 1 should not receive a mark of a point of pattern 2 and vice versa)
(Wiegand and Moloney,2013). So employ the function \code{\link{perm.r1r2.mkdf}} to do
envelope simulation.
}
\examples{
\dontrun{
nsim=2499
ev.mkc.t6.ash.birch.disease.c4<-envelope(ppp.marks.ash.birch.c4,mkcorr.d9.t6,
    funargs=list(m.qli="Species",m.qti="Disease"), nsim=nsim,
    simulate=expression(perm.r1r2.mkdf(ppp.marks.ash.birch.c4,
    m.qli="Species",m.qti="Disease")))
plot(ev.mkc.t6.ash.birch.disease.c4)

ev.mkc.t6.c4<-anylapply(ppp.marks.sp4c.c4,envelope,mkcorr.d9.t6,
    funargs=list(m.qli="Species",m.qti="Disease"), nsim=nsim,
    simulate=simul.bivar.sep)
plot(ev.mkc.t6.c4, main="", main.panel=letters[1:6], legend=FALSE)
}

}
\references{
Wiegand, T. and Moloney, K. A. 2013. Handbook of spatial point-pattern
analysis in ecology. Chapman and Hall/CRC, Boca Raton.

Wiegand, T. 2018. User Manual for the Programita software.
}
\seealso{
\code{\link{mkcorr.d9}}, \code{\link{mkcorr.d7.t6}},
\code{\link{mkcorr.d8.t6}}, \code{\link{mkcorr.t7}} ,\code{\link{mkcorr.t8}}
}
\author{
Tianzhong Jing \email{Jingtianzhong@163.com}
}
