% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkcorr_t7.R
\name{mkcorr.t7}
\alias{mkcorr.t7}
\title{Cumulative Density Correlation Function}
\usage{
mkcorr.t7(X, r=NULL, verbose = TRUE, correction="Ripley",
    method="density",
    weights=NULL,normalise=TRUE, internal=NULL, ...)
}
\arguments{
\item{X}{An object of class "\verb{ppp}"，with one quantitative mark.}

\item{r}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{verbose}{Same as in \verb{localK()} from the package \verb{spatstat}.}

\item{correction}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{method}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{weights}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{normalise}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{internal}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{...}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}
}
\value{
A object of "\verb{fv}".
}
\description{
Estimates cumulative density correlation function for univariate pattern with one quantitative marks.
}
\details{
The density correlation function \eqn{C_{m,K}(r)}is normalized by the product of the
standard deviations \eqn{\sigma_m\sigma_K} of the marks \eqn{m_i} and the individual K-functions \eqn{K_i(r)},
respectively. \emph{C} stands for correlation, \emph{m} for the first mark \eqn{m_i}  and \emph{K}
for the second mark \eqn{K_i(r)} (Fedriani et al., 2015).

The test function (test 7) is
\deqn{C_{m,K}(r)=(m_i-\mu)(K_i(r)-K(r))}
where the function \eqn{K_i()} is the local Ripley K function of each
point \emph{i} and \eqn{K()} is the Ripley K function which equals
to the mean of \eqn{K_i()}, \eqn{m_i} is the mark value of point \emph{i}
and \eqn{\mu} mean mark of all points.

For envelope simulation, user may do a permutation of the mark or a simulation
of a CSR of all points.

This function is similar to the function \eqn{k_{m,g}(r)} which is a non cumulative one.
}
\examples{
\dontrun{
#the data spruces from package spatstat
nsim=29
mkc.t7.spruces<-mkcorr.t7(spruces,correction="translate")
plot(mkc.t7.spruces)
ev.mkc.t7.spruces<-envelope(spruces,mkcorr.t7,
    funargs=list(correction="translate",verbose =FALSE), nsim=nsim,
    simulate=rlabel)
plot(ev.mkc.t7.spruces)
ev.mkc.t7.spruces<-envelope(spruces,mkcorr.t7,
    funargs=list(correction="translate",verbose =FALSE), nsim=nsim)
}

}
\references{
Fedriani, J, T. Wiegand, G. Calvo, A. Suarez-Esteban, M. Jacome, M. Zywiec,
and M.Delibes. 2015. Unravelling conflicting density- and distance-dependent
effects on plant reproduction using a spatially explicit approach.
Journal of Ecology: 103: 1344-1353

Wiegand, T. 2018. User Manual for the Programita software.
}
\seealso{
\code{\link{mkcorr.t8}}, \code{\link{mkcorr.d9.t6}}, \code{\link{mkcorr.d7.t6}},
\code{\link{mkcorr.d8.t6}}, \code{\link{mkcorr.d7.t7}}, \code{\link{mkcorr.d8.t7}}.
}
\author{
Tianzhong Jing \email{Jingtianzhong@163.com}
}
