% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkcorr_t8.R
\name{mkcorr.t8}
\alias{mkcorr.t8}
\title{Density Correlation Function}
\usage{
mkcorr.t8(X, r=NULL, delta=NULL, stoyan=0.15, method="density",
    weights=NULL,normalise=TRUE, internal=NULL,...)
}
\arguments{
\item{X}{An object of class "\verb{ppp}"，with one quantitative mark.}

\item{r}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{delta}{Same as in \verb{localpcf()} from the package \verb{spatstat}.}

\item{stoyan}{Same as in \verb{localpcf()} from the package \verb{spatstat}.}

\item{method}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{weights}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{normalise}{Same as in \verb{markcorr()} from the package \verb{spatstat}}

\item{internal}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}

\item{...}{Same as in \verb{markcorr()} from the package \verb{spatstat}.}
}
\value{
A object of "\verb{fv}".
}
\description{
Estimates density correlation function for univariate pattern with one quantitative marks.
}
\details{
The test function (test 8), is
\deqn{C_{m,g}(r)=(m_i-\mu)(g_i(r)-g(r))}
where the function \eqn{g_i()} is the local pair-correlation function of each
point \emph{i} and \eqn{g()} is the pair-correlation function which equals
to the mean of \eqn{g_i()}, \eqn{m_i} is the mark value of point \emph{i}
and \eqn{\mu} is the mean mark of all points.

For envelope simulation, user may do a permutation of the mark or a simulation
of a CSR of all points.

This function is time consuming.

This function is similar to the function \eqn{C_{m,K}(r)} which is a cumulative one.
}
\examples{
\dontrun{
the data spruces from package spatstat
nsim=2499
mkc.t8.spruces<-mkcorr.t8(spruces)
plot(mkc.t8.spruces)
ev.mkc.t8.spruces<-envelope(spruces,mkcorr.t8, nsim=nsim,simulate=rlabel)
plot(ev.mkc.t8.spruces)
ev.mkc.t8.spruces<-envelope(spruces,mkcorr.t8, nsim=nsim)
}

}
\references{
Fedriani, J, T. Wiegand, G. Calvo, A. Suarez-Esteban, M. Jacome, M. Zywiec,
and M.Delibes. 2015. Unravelling conflicting density- and distance-dependent
effects on plant reproduction using a spatially explicit approach.
Journal of Ecology: 103: 1344-1353

Wiegand, T. 2018. User Manual for the Programita software.
}
\seealso{
\code{\link{mkcorr.t7}}, \code{\link{mkcorr.d9.t6}}, \code{\link{mkcorr.d7.t6}},
\code{\link{mkcorr.d8.t6}}, \code{\link{mkcorr.d7.t8}}, \code{\link{mkcorr.d8.t8}}.
}
\author{
Tianzhong Jing \email{Jingtianzhong@163.com}
}
