% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ND_NDcross.R
\name{ND_NDcross}
\alias{ND_NDcross}
\title{\verb{NDcross} difference estimation, tidy data}
\usage{
ND_NDcross(ppp.marks,yes="Yes",no="No",t.type="yn-nn",
   r = NULL, breaks = NULL, ...,
   k=1, correction = c("rs", "km", "han"))
}
\arguments{
\item{ppp.marks}{A "\verb{ppp}" object with a dichotomous mark (for example, "Yes" and "No").}

\item{yes}{The value of successful category in the dichotomous mark,
for example, "Yes". If \verb{missing}, the second level of the mark will be assigned.}

\item{no}{The value of failure category in the dichotomous mark,
for example, "No". If \verb{missing}, the first level of the mark will be assigned.}

\item{t.type}{Character, indicating what type of test to be estimated. The value can be "yn-nn", "ny-nn", "yy-nn", and "ny-yn",see \verb{Details}}

\item{r}{Same as in \verb{NDcross()}}

\item{breaks}{Same as in \verb{NDcross()}}

\item{...}{Same as in \verb{NDcross()}}

\item{k}{Integer, indicating the \emph{k}th nearest neighbour.}

\item{correction}{Same as in \verb{NDKcross()}}
}
\value{
A object of "\verb{fv}".
}
\description{
Estimates difference between two cross-type \emph{ND} functions (\verb{NDcross}).
}
\details{
Bivariate difference \emph{ND} functions is estimated as (Raventós, et al., 2010).

if t.type=="yn-nn", \deqn{ND_{m,l}(r)-ND_{l,l}(r)}
if t.type=="ny-nn", \deqn{ND_{l,m}(r)-ND_{l,l}(r)}
if t.type=="yy-nn", \deqn{ND_{m,m}(r)-ND_{l,l}(r)}
if t.type=="ny-yn", \deqn{ND_{l,l}(r)-ND_{m,l}(r)}
where \eqn{ND()} is the \emph{k}th nearest neighbor distance function, \emph{l} and \emph{m} are
two levels of a binary mark.

This function tests, for example, the host density-dependent infection of a
forest disease.

Performs the same job as \code{\link{ND_NDcross2}}, but used for tidy data,
i.e, with only one mark.

Use the function \verb{rlabel()} from the package \verb{spatstat} to permute
the binary mark for envelope simulation.
}
\examples{
\dontrun{
plot(ND_NDcross(amacrine,yes="off",no="on",t.type="ny-nn"))
plot(ND_NDcross(ppp.ash.c3,t.type="ny-yn"))
nsim=2499
ev.nd_ndcross.ash.c3<-envelope(ppp.ash.c3,ND_NDcross,nsim=nsim,
    funargs=list(yes="Yes",no="No"),simulate=rlabel)
plot(ev.nd_ndcross.ash.c3)
}
}
\references{
Wiegand, T. and Moloney, K. A. 2013. Handbook of spatial point-pattern
analysis in ecology. Chapman and Hall/CRC, Boca Raton.

Wiegand, T. 2018. User Manual for the Programita software.

Raventós, J., Wiegand, T., and de Luis, M. 2010. Evidence for the spatial
segregation hypothesis: a test with nine-year survivorship data in a
Mediterranean shrubland. Ecology. 91 7:2110-2120.
}
\seealso{
\code{\link{ND_NDcross2}} which does the same job,
\code{\link{g_gcross}}, \code{\link{K_Kcross}}, \code{\link{J_Jcross}},\code{\link{L_Lcross}},
\code{\link{triNDcross}}, \code{\link{triND_ND}}.
}
\author{
Tianzhong Jing \email{Jingtianzhong@163.com}
}
