% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ND_NDcross2.R
\name{ND_NDcross2}
\alias{ND_NDcross2}
\title{\verb{NDcross} difference estimation, untidy data}
\usage{
ND_NDcross2(ppp.marks,V.species="Species", host="Ash", V.bistate="Infected",
    yes="Yes",no="No",t.type="yn-nn",...,
    r = NULL, breaks = NULL,
    k=1, correction = c("rs", "km", "han"))
}
\arguments{
\item{ppp.marks}{A "\verb{ppp}" object with at least two qualitative marks. One is a multiple category mark (for example, species in a forest) and the other is a dichotomous mark (for example, "Yes" and "No").}

\item{V.species}{A column in data represents a multiple category mark, such as tree species in a forest.}

\item{host}{A value of V.species, indicating the focal tree species. That is the host tree of a disease which is concerned.}

\item{V.bistate}{A column in data represents the bi-state of the host tree, for example, infected or uninfected.}

\item{yes}{The value of successful category in the mark \verb{V.bistate}, for example, "Yes". If \verb{missing}, the second level of the mark will be assigned.}

\item{no}{The value of failure category n the mark \verb{V.bistate}, for example, "No". If \verb{missing}, the first level of the mark will be assigned.}

\item{t.type}{Character, indicating what type of test to be estimated. The value can be "yn-nn", "ny-nn", "yy-nn", and "ny-yn",see \verb{Details}}

\item{...}{Same as in \verb{NDcross()}}

\item{r}{Same as in \verb{NDcross()}}

\item{breaks}{Same as in \verb{NDcross()}}

\item{k}{Integer, indicating the \emph{k}th nearest neighbour.}

\item{correction}{Same as in \verb{NDcross()}}
}
\value{
A object of "\verb{fv}".
}
\description{
Estimates difference between two cross-type \emph{ND} functions (\verb{NDcross}).
}
\details{
if t.type=="yn-nn", \deqn{ND_{m,l}(r)-ND_{l,l}(r)}
if t.type=="ny-nn", \deqn{ND_{l,m}(r)-ND_{l,l}(r)}
if t.type=="yy-nn", \deqn{ND_{m,m}(r)-ND_{l,l}(r)}
if t.type=="ny-yn", \deqn{ND_{l,l}(r)-ND_{m,l}(r)}
where \eqn{ND()} is the\emph{k}th nearest neighbor distance function, \emph{l} and \emph{m} are
two levels of a binary mark.

Performs the same job as \code{\link{ND_NDcross}}, but for untidy data,
i.e., the mark is a data frame. The mark variable (for example, \verb{Species})
contains a focal tree species and other tree species.
Another mark variable \verb{V.bistate} is only connected to the focal species.

This function tests, for example, the host density-dependent infection of a
forest disease.

Use the function \code{\link{perm.r1fn.mkdf}} to permute \verb{V.bistate} mark of the
focal species for envelope simulation.
}
\examples{
\dontrun{
plot(ND_NDcross2(ppp.marks.c3,t.type="ny-nn"))
nsim=2499
ev.ND_NDcross2.ash.c3<-envelope(ppp.marks.c3,ND_NDcross2,nsim=nsim,
    funargs=list(V.species="Species", host="Ash", V.bistate="Infected"),
    simulate=expression(simul.shuffle(ppp.marks.c3,V.species="Species",
    host="Ash",V.bistate="Infected")))
plot(ev.ND_NDcross2.ash.c3)
}

}
\references{
Wiegand, T. and Moloney, K. A. 2013. Handbook of spatial point-pattern
analysis in ecology. Chapman and Hall/CRC, Boca Raton.

Wiegand, T. 2018. User Manual for the Programita software.

Raventós, J., Wiegand, T., and de Luis, M. 2010. Evidence for the spatial
segregation hypothesis: a test with nine-year survivorship data in a
Mediterranean shrubland. Ecology. 91 7:2110-2120.
}
\seealso{
\code{\link{ND_NDcross}} which does the same job,
\code{\link{g_gcross}2}, \code{\link{K_Kcross2}}, \code{\link{J_Jcross2}},\code{\link{L_Lcross2}},
\code{\link{triNDcross}}, \code{\link{triND_ND}}.
}
\author{
Tianzhong Jing \email{Jingtianzhong@163.com}
}
