% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ND_NDdot.R
\name{ND_NDdot}
\alias{ND_NDdot}
\title{\verb{NDdot} difference estimation, tidy data}
\usage{
ND_NDdot(ppp.marks,yes="Yes",no="No",...,
    r = NULL, breaks = NULL,
    k=1,correction = c("km", "rs", "han"))
}
\arguments{
\item{ppp.marks}{A "\verb{ppp}" object with a dichotomous mark(for example, "Yes" and "No").}

\item{yes}{The value of successful category in the dichotomous mark, for example, "Yes". If \verb{missing}, the second level of the mark will be assigned.}

\item{no}{The value of failure category in the dichotomous mark, for example, "No". If \verb{missing}, the first level of the mark will be assigned.}

\item{...}{Same as in \verb{NDdot()} from the package \verb{spatstat}}

\item{r}{Same as in \verb{NDdot()} from the package \verb{spatstat}}

\item{breaks}{Same as in \verb{NDdot()} from the package \verb{spatstat}}

\item{k}{Integer, indicating the \emph{k}th nearest neighbor.}

\item{correction}{Same as in \verb{NDdot()} from the package \verb{spatstat}}
}
\value{
A object of "\verb{fv}".
}
\description{
Estimates difference between two dot-type \verb{ND} functions (\verb{NDdot}).
}
\details{
Bivariate difference \emph{ND} functions is estimated as
\deqn{ND_{l,l+m}(r)-ND_{m,l+m}(r)}
where \eqn{ND()} is the kth nearest neighbor distance function, \emph{l} and \emph{m} are
two levels of a binary mark.

This function tests, for example, the host density-dependent infection of a forest disease.


Use the function \verb{rlabel()} from the package \verb{spatstat} to shuffle
the binary mark for envelope simulation.
}
\examples{
\dontrun{
plot(ND_NDdot(ppp.ash.c3))
nsim=2499
ev.nd_nddot.ash.c3<-envelope(ppp.ash.c3,ND_NDdot,nsim=nsim,
    funargs=list(yes="Yes",no="No"),simulate=rlabel)
plot(ev.nd_nddot.ash.c3)
}
}
\references{
Wiegand, T. and Moloney, K. A. 2013. Handbook of spatial point-pattern
analysis in ecology. Chapman and Hall/CRC, Boca Raton.

Raventós, J., Wiegand, T., and de Luis, M. 2010. Evidence for the spatial
segregation hypothesis: a test with nine-year survivorship data in a
Mediterranean shrubland. Ecology. 91 7:2110-2120.
}
\seealso{
\code{\link{ND_NDdot2}} which does the same job,
\code{\link{g_gdot}}, \code{\link{K_Kdot}}, \code{\link{J_Jdot}}, \code{\link{L_Ldot}}
\code{\link{triND_ND}}, \code{\link{triNDcross}}.
}
\author{
Tianzhong Jing \email{Jingtianzhong@163.com}
}
