% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NDcross.R
\name{NDcross}
\alias{NDcross}
\title{Multitype \emph{k}th Nearest Neighbour Distance Function (i-to-j)}
\usage{
NDcross(X, i, j, r = NULL, breaks = NULL, ...,
   k=1, correction = c("rs", "km", "han"))
}
\arguments{
\item{X}{A "\verb{ppp}" object. Same as in \verb{Gcross} from the package \verb{spatstat}}

\item{i}{The type (mark value) of the points in \verb{X }from which distances are measured. Same as in \verb{Gcross} from the package \verb{spatstat}}

\item{j}{The type (mark value) of the points in \verb{X }from which distances are measured. Same as in \verb{Gcross} from the package \verb{spatstat}}

\item{r}{Same as in \verb{Gcross} from the package \verb{spatstat}}

\item{breaks}{Same as in \verb{Gcross} from the package \verb{spatstat}}

\item{...}{Same as in \verb{Gcross} from the package \verb{spatstat}}

\item{k}{Integer, indicating the \emph{k}th nearest neighbour.}

\item{correction}{Same as in \verb{Gcross} from the package \verb{spatstat}}
}
\value{
A object of "\verb{fv}".
}
\description{
For a multitype point pattern, estimate the distribution of the distance from a point of type
\emph{i} to the \emph{k}th nearest point of type \emph{j}.
}
\details{
This function is the same as \verb{Gcross} in \verb{spatatat} package when \verb{k=1}.

Use the function \code{\link{rpoints.r1fn.mkdf}} to simulate a CSR of points corresponding to \verb{yes}
}
\examples{
\dontrun{
plot(NDcross(amacrine,"off","off",k=2))
}

}
\references{
R package \verb{spatstat}
}
\seealso{
\code{\link{NDmulti}}, \code{\link{NDest}},  \code{\link{ND_NDcross}},
\code{\link{ND_NDcross2}},\code{\link{NDdot}},
\code{\link{triNDcross}}.
}
\author{
Tianzhong Jing \email{Jingtianzhong@163.com}
}
