% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NDdot.R
\name{NDdot}
\alias{NDdot}
\title{Multitype \emph{k}th Nearest Neighbour Distance Function (i-to-any)}
\usage{
NDdot(X, i, r = NULL, breaks = NULL, ...,
   k=1,correction = c("km", "rs", "han"))
}
\arguments{
\item{X}{A "\verb{ppp}" object. Same as in \verb{Gdot} from the package \verb{spatstat}}

\item{i}{The type (mark value) of the points in \verb{X }from which distances are measured. Same as in \verb{Gdot} from the package \verb{spatstat}}

\item{r}{Same as in \verb{Gdot} from the package \verb{spatstat}}

\item{breaks}{Same as in \verb{Gdot} from the package \verb{spatstat}}

\item{...}{Same as in \verb{Gdot} from the package \verb{spatstat}}

\item{k}{Integer, indicating the \emph{k}th nearest neighbour.}

\item{correction}{Same as in \verb{Gdot} from the package \verb{spatstat}}
}
\value{
A object of "\verb{fv}".
}
\description{
For a multitype point pattern, estimate the distribution of the distance from a point of type
\emph{i} to the \emph{k}th nearest other point of any type.
}
\details{
This function is the same as \verb{Gdot} in \verb{spatatat} when \verb{k=1}.
}
\examples{
\dontrun{
plot(NDdot(amacrine,"off",k=2))
}

}
\seealso{
\code{\link{NDcross}},\code{\link{NDmulti}},
\code{\link{ND_NDdot}},\code{\link{ND_NDdot2}},
\code{\link{triNDcross}}, \code{\link{NDest}}.
}
\author{
Tianzhong Jing \email{Jingtianzhong@163.com}
}
