% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NDest.R
\name{NDest}
\alias{NDest}
\title{\emph{k}th Neighbour Distance Function \eqn{ND^k(r)}}
\usage{
NDest(X, r = NULL, breaks = NULL, ...,
   k=1, correction = c("rs", "km", "han"), domain = NULL)
}
\arguments{
\item{X}{A object of class "\verb{ppp}", same as that in the funciton \verb{Gest} from the package \verb{spatstat.}}

\item{r}{same as that in the funciton \verb{Gest} from the package \verb{spatstat.}}

\item{breaks}{same as that in the funciton \verb{Gest} from the package \verb{spatstat.}}

\item{...}{same as that in the funciton \verb{Gest} from the package \verb{spatstat.}}

\item{k}{Integer, indicating the \emph{k}th nearest neighbour.}

\item{correction}{same as that in the funciton \verb{Gest} from the package \verb{spatstat.}}

\item{domain}{same as that in the funciton \verb{Gest} from the package \verb{spatstat.}}
}
\value{
A object of "\verb{fv}".
}
\description{
Estimates the \emph{k}th nearest neighbour distance distribution function
\eqn{ND^k(r)} from a point pattern in a window of arbitrary shape.
}
\details{
This function is the same as \verb{Gest} in \verb{spatatat} when \verb{k=1}.
}
\examples{
\dontrun{
# data amacrine is from package spatstat
plot(NDest(amacrine,k=2))
}

}
\references{
R package \verb{spatstat}

Wiegand, T. and Moloney, K. A. 2013. Handbook of spatial point-pattern
analysis in ecology. Chapman and Hall/CRC, Boca Raton.

Wiegand, T. 2018. User Manual for the Programita software.
}
\seealso{
\code{\link{NDcross}}, \code{\link{NDmulti}},
\code{\link{NDdot}},
\code{\link{triNDcross}}.

\code{\link{NDcross}}, \code{\link{NDmulti}},
\code{\link{NDdot}},
\code{\link{triNDcross}}.
}
\author{
Tianzhong Jing <Jingtianzhong@163.com>

Tianzhong Jing \email{Jingtianzhong@163.com}
}
