% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NDmulti.R
\name{NDmulti}
\alias{NDmulti}
\title{Marked \emph{k}th Nearest Neighbour Distance Function}
\usage{
NDmulti(X, I, J, r = NULL, breaks = NULL, ...,
    k=1, disjoint = NULL,
    correction = c("rs", "km", "han"))
}
\arguments{
\item{X}{A "\verb{ppp}" object. Same as in \verb{Gmulti} from the package \verb{spatstat}}

\item{I}{Subset of points of X from which distances are measured. Same as in \verb{Gmulti} from the package \verb{spatstat}}

\item{J}{Subset of points in X to which distances are measured. Same as in \verb{Gmulti} from the package \verb{spatstat}}

\item{r}{Same as in \verb{Gmulti} from the package \verb{spatstat}}

\item{breaks}{Same as in \verb{Gmulti} from the package \verb{spatstat}}

\item{...}{Same as in \verb{Gmulti} from the package \verb{spatstat}}

\item{k}{Integer, indicating the \emph{k}th nearest neighbour.}

\item{disjoint}{Same as in \verb{Gmulti} from the package \verb{spatstat}}

\item{correction}{Same as in \verb{Gmulti} from the package \verb{spatstat}}
}
\value{
An object of class "\verb{fv}"
}
\description{
For a marked point pattern, estimate the distribution of the distance from a typical point in subset \emph{I}
to the \emph{k}th nearest point of subset \emph{J}
}
\details{
This function is the same as \verb{Gmulti} in \verb{spatatat} when \verb{k=1}.
}
\examples{
\dontrun{
# data amacrine is from package spatstat
plot(NDmulti(amacrine, marks(amacrine)=="off", marks(amacrine)=="on"))
plot(NDmulti(amacrine, marks(amacrine)=="off", marks(amacrine)=="on",k=2))
}

}
\references{
R package \verb{spatstat}
}
\seealso{
\code{\link{NDcross}}, \code{\link{ND_NDcross}},
\code{\link{ND_NDcross2}},\code{\link{ND_NDdot}},\code{\link{NDoverNDdot}},
\code{\link{triND_ND}},\code{\link{triNDcross}}.
}
\author{
Tianzhong Jing \email{Jingtianzhong@163.com}
}
