% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm_onemkvar.R
\name{perm.onemkvar}
\alias{perm.onemkvar}
\title{Permutation of a mark variable}
\usage{
perm.onemkvar(ppp.marks,V.perm,...)
}
\arguments{
\item{ppp.marks}{A marked "\verb{ppp}" object with more than one mark.}

\item{V.perm}{The variable for permutation}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A "\verb{ppp}" object.
}
\description{
Permuting a mark variable. In the case of more than one mark variable,
only the given variable is permuted.
}
\examples{
\dontrun{
perm.onemkvar(ppp.ash.infect.dbh.c3,V.perm="DBH")
}

}
\seealso{
\code{\link{perm.r1r2.mkdf}}, \code{\link{perm.r1fn.mkdf}},
\code{\link{rpoints.r1fn.mkdf}}, \code{\link{rpoints.r1f2.1mk}}, \code{\link{simul.shift.r1f2.1mk}},
 \code{\link{simul.shift.r1fn.mkdf}}, \code{\link{simul.jitter.r1f2.1mk}}, \code{\link{simul.jitter.r1fn.mkdf}},
  \code{\link{mkcorr.d8.t6}},\code{\link{mkcorr.d8}}.
}
\author{
Tianzhong Jing \email{Jingtianzhong@163.com}
}
