% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm_r1fn_mkdf.R
\name{perm.r1fn.mkdf}
\alias{perm.r1fn.mkdf}
\title{Permuting a part of data of a variable}
\usage{
perm.r1fn.mkdf(ppp.marks,V.species="Species",host="Ash",
    V.bistate="Infected", ...)
}
\arguments{
\item{ppp.marks}{A "\verb{ppp}" object with at least two mark variable and one of them is qualitative.}

\item{V.species}{A qualitative variable which one level is given by host}

\item{host}{One of category of V.species}

\item{V.bistate}{A mark variable for permutation. Often qualitative (binary), but quantitative is also OK.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A "\verb{ppp}" object
}
\description{
Randomizes partial data of a mark variable corresponding to one of levels of
another qualitative variable
}
\details{
Randomizes the binary label of a host species.
Often used in envelope simulation of trivariate random labeling.
}
\examples{
\dontrun{
perm.r1fn.mkdf(ppp.marks.c3,V.species="Species",host="Ash",V.bistate="Infected")
}

}
\seealso{
\code{\link{perm.r1r2.mkdf}}, \code{\link{perm.onemkvar}},
\code{\link{rpoints.r1fn.mkdf}}, \code{\link{rpoints.r1f2.1mk}}, \code{\link{simul.shift.r1f2.1mk}},
 \code{\link{simul.shift.r1fn.mkdf}}, \code{\link{simul.jitter.r1f2.1mk}}, \code{\link{simul.jitter.r1fn.mkdf}},
  \code{\link{g_gcross2}}, \code{\link{K_Kcross2}},
\code{\link{g_gdot2}}, \code{\link{K_Kdot2}}, \code{\link{govergdot2}},
\code{\link{KoverKdot2}}.
}
\author{
Tianzhong Jing \email{Jingtianzhong@163.com}
}
