% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm_r1r2_mkdf.R
\name{perm.r1r2.mkdf}
\alias{perm.r1r2.mkdf}
\title{Permuting two parts of a mark separately}
\usage{
perm.r1r2.mkdf(ppp.marks,m.qli="Species",sp1="Ash", sp2="Birch",
    m.qti="Disease",...)
}
\arguments{
\item{ppp.marks}{A "\verb{ppp}" object with at least two mark variable and one of them is binary.}

\item{m.qli}{A qualitative mark variable. If there is more than two categories, only the first two levels are considered.}

\item{sp1}{A mark value of \verb{m.qli}, if \verb{missing}, the first level of \verb{m.qli}}

\item{sp2}{Another mark value of \verb{m.qli}, if \verb{missing}, the second level of \verb{m.qli}}

\item{m.qti}{A mark variable of \verb{ppp.marks}, usually quantitative.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A "\verb{ppp}" object.
}
\description{
For a mark dataframe, permuting two parts of a mark separately corresponding to two levels of another mark variable
}
\details{
This function permutates \verb{m.qti} corresponding to \verb{sp1} and \verb{sp2} seperately.

Given a category variable "Species" has two levels (Ash, Birch) and another mark variable "DBH",
i.e., the sizes of tree,  this function randomizes the size of each species separately.
For other levels of variable "Species",for example, Cork, their sizes keep unchanged.

This function may be used for envelope simulation for data 9.
}
\examples{
\dontrun{
#perm.r1r2.mkdf(ppp.ash.infect.dbh.c3,m.qli="Infected",m.qti="DBH")
}

}
\seealso{
\code{\link{perm.onemkvar}}, \code{\link{perm.r1fn.mkdf}},
\code{\link{rpoints.r1fn.mkdf}}, \code{\link{rpoints.r1f2.1mk}}, \code{\link{simul.shift.r1f2.1mk}},
\code{\link{simul.shift.r1fn.mkdf}}, \code{\link{simul.jitter.r1f2.1mk}}, \code{\link{simul.jitter.r1fn.mkdf}},
\code{\link{mkcorr.d9.t6}}, \code{\link{mkcorr.d9}}
}
\author{
Tianzhong Jing \email{Jingtianzhong@163.com}
}
