% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppp.ash.c3.R
\docType{data}
\name{ppp.ash.c3}
\alias{ppp.ash.c3}
\title{An example of multitype \verb{ppp} object}
\format{
identical \verb{ppp} object in the package \pkg{spatstat}
}
\description{
A marked spatial point pattern with a binary qualitative mark, for example,
\sQuote{Yes} and \sQuote{No}.
}
\details{
Built-in data in markSPPA
}
\keyword{dataset}
