% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppp.ash.g1.R
\docType{data}
\name{ppp.ash.g1}
\alias{ppp.ash.g1}
\title{"ppp" objects for test}
\format{
identical "ppp" in the pagckae spatstat
}
\description{
A marked spatial point pattern with two quantitative mark
}
\details{
Built-in data in markSPPA
}
\keyword{dataset}
