% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppp.ash.infect.dbh.c3.R
\docType{data}
\name{ppp.ash.infect.dbh.c3}
\alias{ppp.ash.infect.dbh.c3}
\title{"ppp" objects for test}
\format{
identical "ppp" in the pagckae spatstat
}
\description{
A marked spatial point pattern with one binary qualitative mark (Infection:
Yes or No) and one quantitative mark(DBH)
}
\details{
Built-in data in markSPPA
}
\keyword{dataset}
