% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppp.marks.c3.R
\docType{data}
\name{ppp.marks.c3}
\alias{ppp.marks.c3}
\title{"ppp" objects for test}
\format{
identical "ppp" in the pagckae spatstat
}
\description{
A marked spatial point pattern with more than one mark. Untidy data.
}
\details{
Built-in data in markSPPA
}
\keyword{dataset}
