% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppp.marks.sp4c.c4.R
\docType{data}
\name{ppp.marks.sp4c.c4}
\alias{ppp.marks.sp4c.c4}
\title{"ppp" objects for test}
\format{
identical "ppp" in the pagckae spatstat
}
\description{
A "ppp" list for anylapply() in the package spatstat.
}
\details{
Built-in data in markSPPA
}
\keyword{dataset}
