% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpoints_r1f2_1mk.R
\name{rpoints.r1f2.1mk}
\alias{rpoints.r1f2.1mk}
\title{Random simulation of the points corresponding to one level of a qualitative mark}
\usage{
rpoints.r1f2.1mk(ppp.marks,p.type="CSR",yes="Yes",...)
}
\arguments{
\item{ppp.marks}{A "\verb{ppp}" object with a multitype qualitative mark variable}

\item{p.type}{Indicating what type of points to be generated, see \verb{Details}.}

\item{yes}{The level of which points to be randomized. If \verb{missing}, the second level of the mark will be assigned.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A "\verb{ppp}" object.
}
\description{
Given a "ppp" object with a multitype qualitative mark variable, this function
does a random simulation of the points corresponding to one level of the mark.
}
\details{
Two type of points are supported, including "CSR" and
"in.unif", which generateed complete spatial random (CSR) points and
independent uniform random points, respectively. For "CSR", the rows of simulated marks
were sampled randomly from the mark data frame corresponding to "Ash"; for "in.unif",
the order of the rows of simulated marks did not change.
}
\examples{
\dontrun{
rpoints.r1f2.1mk(ppp.ash.c3,yes="Yes")
}

}
\seealso{
\code{\link{perm.r1r2.mkdf}}, \code{\link{perm.onemkvar}}, \code{\link{perm.r1fn.mkdf}},
\code{\link{rpoints.r1fn.mkdf}}, \code{\link{simul.shift.r1f2.1mk}},
 \code{\link{simul.shift.r1fn.mkdf}}, \code{\link{simul.jitter.r1f2.1mk}}, \code{\link{simul.jitter.r1fn.mkdf}},
  \code{\link{heterotrl}}, \code{\link{contrl2}},
\code{\link{trig_g}},\code{\link{triK_K}}.
}
\author{
Tianzhong Jing \email{Jingtianzhong@163.com}
}
