% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpoints_r1fn_mkdf.R
\name{rpoints.r1fn.mkdf}
\alias{rpoints.r1fn.mkdf}
\title{Random simulation of the points corresponding to one level of a qualitative mark}
\usage{
rpoints.r1fn.mkdf(ppp.marks,p.type="CSR",V.species="Species",sp="Cork",...)
}
\arguments{
\item{ppp.marks}{A "\verb{ppp}" object with a mark dataframe including at least one qualitative variable}

\item{p.type}{Indicating what type of points to be generated, see \verb{Details}.}

\item{V.species}{The name of a qualitative mark variable}

\item{sp}{The level of which points to be simulated. If missing, the first level of \verb{V.species}}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A "\verb{ppp}" object.
}
\description{
Given a "ppp" object with a  mark dataframe, this function
does a random simulation of the points corresponding to one level of a qualitative mark,
randomly assigns corresponding mark rows,
and keeps the points corresponding to other levels unchanged.
}
\details{
Two type of points are supported, including "CSR" and
"in.unif", which generateed complete spatial random (CSR) points and
independent uniform random points, respectively. For "CSR", the rows of simulated marks
were sampled randomly from the mark data frame corresponding to "Ash"; for "in.unif",
the order of the rows of simulated marks did not change.
}
\examples{
\dontrun{
rpoints.r1fn.mkdf(ppp.marks.c3,V.species="Species",sp="Cork")
}

}
\seealso{
\code{\link{perm.r1r2.mkdf}}, \code{\link{perm.onemkvar}},
\code{\link{rpoints.r1fn.mkdf}}, \code{\link{rpoints.r1f2.1mk}}, \code{\link{simul.shift.r1f2.1mk}},
 \code{\link{simul.shift.r1fn.mkdf}}, \code{\link{simul.jitter.r1f2.1mk}}, \code{\link{simul.jitter.r1fn.mkdf}},
  \code{\link{heterotrl}}, \code{\link{contrl2}},
\code{\link{trig_g}},\code{\link{triK_K}}.
}
\author{
Tianzhong Jing \email{Jingtianzhong@163.com}
}
