% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul_jitter_r1f2_1mk.R
\name{simul.jitter.r1f2.1mk}
\alias{simul.jitter.r1f2.1mk}
\title{Independently randomly displaces points corresponding to one level of a qualitative mark}
\usage{
simul.jitter.r1f2.1mk(ppp.marks,yes="Yes",radius,
   retry=TRUE, giveup = 10000, trim=FALSE,
   ..., nsim=1, drop=TRUE)
}
\arguments{
\item{ppp.marks}{A "\verb{ppp}" object with a multitype qualitative mark variable}

\item{yes}{The level of which points to be displaced. If \verb{missing}, the second level of the mark will be assigned.}

\item{radius}{Same as in \verb{rjitter()} from the package \verb{spatstat}}

\item{retry}{Same as in \verb{rjitter()} from the package \verb{spatstat}}

\item{giveup}{Same as in \verb{rjitter()} from the package \verb{spatstat}}

\item{trim}{Same as in \verb{rjitter()} from the package \verb{spatstat}}

\item{...}{Same as in \verb{rjitter()} from the package \verb{spatstat}.}

\item{nsim}{Same as in \verb{rjitter()} from the package \verb{spatstat}}

\item{drop}{Same as in \verb{rjitter()} from the package \verb{spatstat}}
}
\value{
A "\verb{ppp}" object.
}
\description{
Given a "ppp" object with a multitype qualitative mark variable, this function
does an independent random displacement of the points corresponding to one level of the mark.
}
\details{
This function activates the function \verb{rshift} from the package \verb{spatstat}.
For a multitype ppp, this function only displaces the points corresponding to pattern \verb{yes}.
}
\examples{
\dontrun{
simul.jitter.r1f2.1mk(ppp.ash.c3,yes="Yes",radius=0.1)
}

}
\seealso{
\code{\link{perm.r1r2.mkdf}}, \code{\link{perm.onemkvar}}, \code{\link{perm.r1fn.mkdf}},
\code{\link{rpoints.r1fn.mkdf}}, \code{\link{rpoints.r1f2.1mk}}, \code{\link{simul.shift.r1f2.1mk}},
 \code{\link{simul.shift.r1fn.mkdf}}, \code{\link{simul.jitter.r1fn.mkdf}},
}
\author{
Tianzhong Jing \email{Jingtianzhong@163.com}
}
