% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul_jitter_r1fn_mkdf.R
\name{simul.jitter.r1fn.mkdf}
\alias{simul.jitter.r1fn.mkdf}
\title{Random simulation of the points corresponding to one level of a qualitative mark}
\usage{
simul.jitter.r1fn.mkdf(ppp.marks,V.species="Species",sp="Cork",radius,
   retry=TRUE, giveup = 10000, trim=FALSE,
   ..., nsim=1, drop=TRUE)
}
\arguments{
\item{ppp.marks}{A "\verb{ppp}" object with a mark dataframe including at least one qualitative variable}

\item{V.species}{The name of a qualitative mark variable}

\item{sp}{The level of which points to be simulated. If \verb{missing}, the first level of \verb{V.species}}

\item{radius}{Same as in \verb{rjitter()} from the package \verb{spatstat}}

\item{retry}{Same as in \verb{rjitter()} from the package \verb{spatstat}}

\item{giveup}{Same as in \verb{rjitter()} from the package \verb{spatstat}}

\item{trim}{Same as in \verb{rjitter()} from the package \verb{spatstat}}

\item{...}{Same as in \verb{rjitter()} from the package \verb{spatstat}.}

\item{nsim}{Same as in \verb{rjitter()} from the package \verb{spatstat}}

\item{drop}{Same as in \verb{rjitter()} from the package \verb{spatstat}}
}
\value{
A "\verb{ppp}" object.
}
\description{
Given a "ppp" object with a  mark dataframe, this function
does a random simulation of the points corresponding to one level of a qualitative mark,
randomly assigns corresponding mark rows,
and keeps the points corresponding to other levels unchanged.
}
\details{
Two type of points are supported, including "CSR" and
"in.unif", which generateed complete spatial random (CSR) points and
independent uniform random points, respectively. For "CSR", the rows of simulated marks
were sampled randomly from the mark data frame corresponding to "Ash"; for "in.unif",
the order of the rows of simulated marks did not change.
}
\examples{
\dontrun{
simul.jitter.r1fn.mkdf(ppp.marks.c3,V.species="Species",sp="Cork",radius=0.1)
}

}
\seealso{
\code{\link{perm.r1r2.mkdf}}, \code{\link{perm.onemkvar}}, \code{\link{perm.r1fn.mkdf}},
\code{\link{rpoints.r1fn.mkdf}}, \code{\link{rpoints.r1f2.1mk}}, \code{\link{simul.shift.r1f2.1mk}},
\code{\link{simul.shift.r1fn.mkdf}}, \code{\link{simul.jitter.r1f2.1mk}},
\code{\link{heterotrl}}, \code{\link{contrl2}},
\code{\link{trig_g}},\code{\link{triK_K}}.
}
\author{
Tianzhong Jing \email{Jingtianzhong@163.com}
}
