% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul_shift_r1f2_1mk.R
\name{simul.shift.r1f2.1mk}
\alias{simul.shift.r1f2.1mk}
\title{Randomly shifts the points corresponding to one level of a qualitative mark}
\usage{
simul.shift.r1f2.1mk(ppp.marks,yes="Yes",...,
    which=NULL, group, nsim=1, drop=TRUE)
}
\arguments{
\item{ppp.marks}{A "\verb{ppp}" object with a multitype qualitative mark variable}

\item{yes}{The level of which points to be randomized. If \verb{missing}, the second level of the mark will be assigned.}

\item{...}{Same as in \verb{rshift()} from the package \verb{spatstat}.}

\item{which}{Same as in \verb{rshift()} from the package \verb{spatstat}.}

\item{group}{Same as in \verb{rshift()} from the package \verb{spatstat}.}

\item{nsim}{Same as in \verb{rshift()} from the package \verb{spatstat}.}

\item{drop}{Same as in \verb{rshift()} from the package \verb{spatstat}.}
}
\value{
A "\verb{ppp}" object.
}
\description{
Given a "ppp" object with a multitype qualitative mark variable, this function
does a shift of the points corresponding to one level of the mark.
}
\details{
This function activates the function \verb{rshift} from the package \verb{spatstat}.
For a multitype ppp, this function only shifts the points corresponding to pattern \verb{yes}.
}
\examples{
\dontrun{
simul.shift.r1f2.1mk(amacrine,radius=0.1,edge="erode")
}

}
\seealso{
\code{\link{perm.r1r2.mkdf}}, \code{\link{perm.onemkvar}}, \code{\link{perm.r1fn.mkdf}},
\code{\link{rpoints.r1fn.mkdf}}, \code{\link{rpoints.r1f2.1mk}},
 \code{\link{simul.shift.r1fn.mkdf}}, \code{\link{simul.jitter.r1f2.1mk}}, \code{\link{simul.jitter.r1fn.mkdf}},
\code{\link{trig_g}},\code{\link{triK_K}}, \code{\link{heterotrl}}, \code{\link{contrl2}}.
}
\author{
Tianzhong Jing \email{Jingtianzhong@163.com}
}
