% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trig_g.R
\name{trig_g}
\alias{trig_g}
\title{Trivariate type \verb{pcfcross} difference estimation}
\usage{
trig_g(ppp.marks,V.species="Species",host="Ash", nonhost="Cork",
    V.bistate="Infected",yes="Yes",no="No",...,
    r = NULL,
    kernel = "epanechnikov", bw = NULL, stoyan = 0.15,
    correction = c("isotropic", "Ripley", "translate"),
    divisor = c("r", "d"),
    ratio = FALSE)
}
\arguments{
\item{ppp.marks}{A marked "\verb{ppp}" object with two qualitative marks, one is multiple and the other is binary.}

\item{V.species}{The multiple qualitative mark variable}

\item{host}{One of levels of \verb{V.species}, for example, a host tree species. If \verb{missing}, the 1st level of \verb{V.species} will be assigned.}

\item{nonhost}{One of levels of \verb{V.species}, for example, a nonhost tree species. If \verb{missing}, the 2nd level of \verb{V.species} will be assigned.}

\item{V.bistate}{A binary mark only connected to \verb{host}.}

\item{yes}{The value of successful category in the mark \verb{V.bistate}, for example, "Yes". If \verb{missing}, the second level of the mark will be assigned.}

\item{no}{The value of failure category n the mark \verb{V.bistate}, for example, "No". If \verb{missing}, the first level of the mark will be assigned.}

\item{...}{Same as in as in \verb{pcfcross()} from the package \verb{spatstat}}

\item{r}{Same as in \verb{pcfcross()}from the package \verb{spatstat}}

\item{kernel}{Same as in \verb{pcfcross()} from the package \verb{spatstat}}

\item{bw}{Same as in \verb{pcfcross()} from the package \verb{spatstat}}

\item{stoyan}{Same as in \verb{pcfcross()} from the package \verb{spatstat}}

\item{correction}{Same as in \verb{pcfcross()} from the package \verb{spatstat}}

\item{divisor}{Same as in \verb{pcfcross()} from the package \verb{spatstat}}

\item{ratio}{Same as in \verb{pcfcross()} from the package \verb{spatstat}}
}
\value{
A "\verb{fv}" object
}
\description{
Estimates difference between two trivariate type pcfcross functions
}
\details{
Trivariate difference K functions is estimated as (Raventós, et al., 2010).
\deqn{g_{l,a}(r)-g_{m,a}(r)}
where \eqn{g()} is the pair-correlation function, \emph{l} and \emph{m} are
two levels of a binary mark of the focal pattern, \emph{a} represents an antecedent pattern.

For example, this function tests whether the infected host trees have more neighboring nonhost trees than uninfected host trees.

Use the function \code{\link{rpoints.r1fn.mkdf}} to simulate a CSR pattern of the
antecedent pattern for envelope simulation, or use the function \code{\link{perm.r1fn.mkdf}} to
permute \verb{V.bistate} mark of the focal species.
}
\examples{
\dontrun{
plot(trig_g(ppp.marks.c3))
nsim=2499
ev.trig_g.cork.c3<-envelope(ppp.marks.c3,trig_g,nsim=nsim,
    funargs=list(V.species="Species", host="Ash",nonhost="Cork"),
    simulate = expression(rpoints.r1fn.mkdf(ppp.marks.c3,V.species="Species",sp="Cork")))
plot(ev.trig_g.cork.c3)
}

}
\references{
Wiegand, T. and Moloney, K. A. 2013. Handbook of spatial point-pattern
analysis in ecology. Chapman and Hall/CRC, Boca Raton.

Raventós, J., Wiegand, T., and de Luis, M. 2010. Evidence for the spatial
segregation hypothesis: a test with nine-year survivorship data in a
Mediterranean shrubland. Ecology. 91 7:2110-2120.
}
\seealso{
\code{\link{triK_K}},\code{\link{g_gcross}},\code{\link{g_gdot}},
\code{\link{govergdot}}, \code{\link{trigcross}}.
}
\author{
Tianzhong Jing \email{Jingtianzhong@163.com}
}
