% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigcross.R
\name{trigcross}
\alias{trigcross}
\title{Trivariate \verb{Kcross}}
\usage{
trigcross(ppp.marks,V.species="Species",host="Ash", nonhost="Cork",
   V.bistate="Infected",yes="Yes",...,
   r = NULL,
   kernel = "epanechnikov", bw = NULL, stoyan = 0.15,
   correction = c("isotropic", "Ripley", "translate"),
   divisor = c("r", "d"),
   ratio = FALSE)
}
\arguments{
\item{ppp.marks}{A "\verb{ppp}" object with at least two qualitative marks. One is a multiple category mark (for example, species in a forest) and the other is a dichotomous mark (for example, "Yes" and "No").}

\item{V.species}{A column in data represents a multiple category mark, such as tree species in a forest.}

\item{host}{A value of V.species, indicating the focal tree species. That is the host tree of a disease which is concerned.If \verb{missing}, the 1st level of \verb{V.species} will be assigned.}

\item{nonhost}{One of levels of \verb{V.species}, for example, a nonhost tree species. If \verb{missing}, the 2nd level of \verb{V.species} will be assigned.}

\item{V.bistate}{A column in data represents the bi-state of the host tree, for example, infected or uninfected.}

\item{yes}{A value indicates the state of the host tree of concerned, for example, "Yes" or "No". If \verb{missing}, the second level of \verb{V.bistate } will be assigned.}

\item{...}{Same as in \verb{pcfcross()} from the package \verb{spatstat}}

\item{r}{Same as in \verb{pcfcross()}from the package \verb{spatstat}}

\item{kernel}{Same as in \verb{pcfcross()} from the package \verb{spatstat}}

\item{bw}{Same as in \verb{pcfcross()} from the package \verb{spatstat}}

\item{stoyan}{Same as in \verb{pcfcross()} from the package \verb{spatstat}}

\item{correction}{Same as in \verb{pcfcross()} from the package \verb{spatstat}}

\item{divisor}{Same as in \verb{pcfcross()} from the package \verb{spatstat}}

\item{ratio}{Same as in \verb{pcfcross()} from the package \verb{spatstat}}
}
\value{
A object of "\verb{fv}".
}
\description{
Estimates trivariate cross-type K function (Kcross).
}
\details{
The trivariate cross type  \eqn{g_{l,h}(r)},where h represents a heterospecies,
was extended from bivariate cross type \eqn{g_{l,m}(r)}

The mark is a data frame. A mark variable (for example, \verb{Species})
contains a focal tree species and other tree species.
Another mark variable \verb{V.bistate} is only connected to the focal species.

This function tests, for example, the infected host trees aggregated/segregated with a nonhost species.

Use the function \code{\link{perm.r1fn.mkdf}} to permute \verb{V.bistate} mark of the
focal species for envelope simulation, or use \code{\link{rpoints.r1fn.mkdf}} to generate
a CSR of \verb{nonhost}.
}
\examples{
\dontrun{
plot(trigcross(ppp.marks.c3),ylim=c(0,2))
nsim=2499
ev.trigcross.cork.c3<-envelope(ppp.marks.c3,trigcross,nsim=nsim,
   funargs=list(V.species = "Species",host="Ash",nonhost="Cork",V.bistate="Infected",yes="Yes"),
   simulate = expression(rpoint.r1fn.mk(ppp.marks.c3,V.species="Species",sp="Cork")))
plot(ev.trigcross.cork.c3)
}

}
\references{
Wiegand, T. and Moloney, K. A. 2013. Handbook of spatial point-pattern
analysis in ecology. Chapman and Hall/CRC, Boca Raton.

Raventós, J., Wiegand, T., and de Luis, M. 2010. Evidence for the spatial
segregation hypothesis: a test with nine-year survivorship data in a
Mediterranean shrubland. Ecology. 91 7:2110-2120.
}
\seealso{
\code{\link{triLcross}},\code{\link{g_gcross}}, \code{\link{K_Kcross2}},
\code{\link{g_gcross2}},\code{\link{g_gdot2}},\code{\link{govergdot2}},
\code{\link{trig_g}},\code{\link{triKcross}}.
}
\author{
Tianzhong Jing \email{Jingtianzhong@163.com}
}
