% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triND_ND.R
\name{triND_ND}
\alias{triND_ND}
\title{Trivariate type \verb{Kcross} difference estimation}
\usage{
triND_ND(ppp.marks,V.species="Species",host="Ash", nonhost="Cork",
    V.bistate="Infected",yes="Yes",no="No",...,
    r = NULL, breaks = NULL,
    k=1, correction = c("rs", "km", "han"))
}
\arguments{
\item{ppp.marks}{A marked "\verb{ppp}" object with two qualitative marks, one is multiple and the other is binary.}

\item{V.species}{The multiple qualitative mark variable}

\item{host}{One of levels of \verb{V.species}, for example, a host tree species. If \verb{missing}, the 1st level of \verb{V.species} will be assigned.}

\item{nonhost}{One of levels of \verb{V.species}, for example, a nonhost tree species. If \verb{missing}, the 2nd level of \verb{V.species} will be assigned.}

\item{V.bistate}{A binary mark only connected to \verb{host}.}

\item{yes}{The value of successful category in the mark \verb{V.bistate}, for example, "Yes". If \verb{missing}, the second level of the mark will be assigned.}

\item{no}{The value of failure category n the mark \verb{V.bistate}, for example, "No". If \verb{missing}, the first level of the mark will be assigned.}

\item{...}{Same as in as in \verb{Kcross()} from the package \verb{spatstat}}

\item{r}{Same as in \verb{Kcross()} from the package \verb{spatstat}}

\item{breaks}{Same as in \verb{Kcross()} from the package \verb{spatstat}}

\item{k}{Integer, indicating the \emph{k}th nearest neighbour.}

\item{correction}{Same as in \verb{Kcross()} from the package \verb{spatstat}}
}
\value{
An object of "\verb{fv}".
}
\description{
Estimates difference between two trivariate type Kcross functions
}
\details{
Trivariate difference K functions is estimated as (Raventós, et al., 2010).
\deqn{ND_{l,a}(r)-ND_{m,a}(r)}
where \eqn{ND()} is the \verb{NDcross} function, \emph{l} and \emph{m} are
two levels of a binary mark of the focal pattern, \emph{a} represents an antecedent pattern.

For example, this function tests whether the infected host trees have more neighboring nonhost trees than uninfected host trees.

Use the function \code{\link{rpoints.r1fn.mkdf}} to simulate a CSR pattern of the
antecedent pattern for envelope simulation, or use the function \code{\link{perm.r1fn.mkdf}} to
permute \verb{V.bistate} mark of the focal species.
}
\examples{
\dontrun{
plot(triND_ND(ppp.marks.c3))
nsim=2499
ev.trind_nd.cork.c3<-envelope(ppp.marks.c3,triND_ND,nsim=nsim,
    funargs=list(V.species="Species", host="Ash",nonhost="Cork"),
    simulate = expression(rpoints.r1fn.mkdf(ppp.marks.c3,V.species="Species",sp="Cork")))
plot(ev.trind_nd.cork.c3)
}

}
\references{
Wiegand, T. and Moloney, K. A. 2013. Handbook of spatial point-pattern
analysis in ecology. Chapman and Hall/CRC, Boca Raton.

Raventós, J., Wiegand, T., and de Luis, M. 2010. Evidence for the spatial
segregation hypothesis: a test with nine-year survivorship data in a
Mediterranean shrubland. Ecology. 91 7:2110-2120.
}
\seealso{
\code{\link{triNDcross}},\code{\link{NDcross}},\code{\link{triK_K}},
\code{\link{triL_L}},\code{\link{trig_g}}.
}
\author{
Tianzhong Jing \email{Jingtianzhong@163.com}
}
